==> 04-scripts/01-data-cleaning.qmd <==
---
title: "Data Cleaning"
format: html
editor: visual
editor_options: 
  chunk_output_type: console
---

# Data Cleaning and Setup

## EMA

```{r}
ema_df <- read_csv("../02-data/01-raw-data/erdyn-ema-2025-06-13.csv")

ema_items <- cb %>% filter(level == "EMA" & category != "proc" & !is.na(item)) 

ema_long <- ema_df %>%
  pivot_longer(
    cols = ema_items$item_name
    , names_to = "item_name"
    , values_to = "value"
  ) %>% 
  left_join(
    ema_items %>%
      select(item, item_name, category)
    )
```

### Emotions

```{r}
pomp <- function(x, mini = 1, maxi = 5) ((x - mini)/(maxi- mini))*10

emo_long <- ema_long %>%
  filter(category == "emotion") %>%
  mutate(value = pomp(value))

emo_long <- emo_long %>%
  group_by(SID, Date, StartDate, Day, Hour, HourBlock, sample, item) %>%
  summarize(value = mean(value, na.rm = T)) %>%
  ungroup() %>%
  full_join(
    emo_long %>%
      filter(category == "emotion") %>%
      select(-category, -item) %>%
      rename(item = item_name)
  )
```

### Emotion Regulation Goals

```{r}
ergoals_long <- ema_long %>%
  filter(category == "emotion regulation goals") %>%
  mutate(value = pomp(value, -2, 2)-5)
```

### Emotion Regulation Strategies

```{r}
score <- function(x) {y <- if(any(x == 1)) 1 else 0}
erstrat_long <- ema_long %>%
  filter(category == "emotion regulation strategies") %>%
  group_by(SID, sample, StartDate, item_name) %>%
  summarize(value = score(value)) %>% 
  group_by(SID, sample, StartDate) %>%
  summarize(er_strat_rep = sum(value, na.rm = T)) %>%
  ungroup()

ema_strat_use <- ema_long %>%
  filter(category == "emotion regulation strategies") %>%
  group_by(SID, sample, StartDate, item_name, Date) %>%
  summarize(value = score(value)) %>% 
  group_by(SID, sample, StartDate, Date) %>%
  summarize(er_strat_rep = score(value))  %>%
  ungroup()

ergoals_long <- ergoals_long %>%
  full_join(ema_strat_use) %>%
  mutate(er_use = ifelse(value == 0 & er_strat_rep == 0, 0, 1)) 

ergoals_long %>%
  group_by(SID, item_name, er_use) %>%
  tally() %>%
  mutate(perc = n/sum(n)*100) %>%
  filter(er_use == 1) %>%
  ggplot(aes(x = perc)) + 
  geom_histogram(fill = "blue", alpha = .5, color = "black") + 
  facet_wrap(~item_name) + 
  my_theme()

ergoals_long %>%
  full_join(ema_strat_use) %>%
  mutate(er_use = ifelse(value == 0 & er_strat_rep == 0, 0, 1)) %>%
  filter(er_use == 1)

ema_strat_use %>%
  group_by(SID, er_strat_rep) %>%
  tally() %>%
  mutate(perc = n/sum(n)*100) %>%
  filter(er_strat_rep == 1) %>%
  ggplot(aes(x = perc)) + 
  geom_histogram(fill = "blue", alpha = .5, color = "black") + 
  my_theme()
## add in maintain + no strategies
  

erstrat_long %>%
  ggplot(aes(x = er_strat_rep, fill = sample)) + 
    geom_histogram(position = "dodge", color = "black", alpha = .5) + 
    scale_fill_manual(values = c("blue", "pink3")) + 
    theme_classic()


erstrat_avg <- ema_long %>%
  filter(category == "emotion regulation strategies") %>%
  group_by(SID, StartDate, item_name)%>%
  summarize(avg_use = mean(value), .groups = "drop") %>%
  pivot_wider(names_from = item_name, values_from = avg_use, names_prefix = "avg_")


erstrat<-left_join(erstrat_long,erstrat_avg)


```

### Recompile

#### RQ1

```{r}
rescale_time <- function(x){
  x %>%
    arrange(Date) %>%
    mutate(time = difftime(Date, Date[1], unit = "hours")) %>%
    as.data.frame()
}

nested_rq1 <- emo_long %>%
  rename(emo_item = item, emo_value = value) %>%
  drop_na(emo_item, emo_value, Date) %>%
  mutate(SID2 = SID) %>%
  group_by(SID2, sample, emo_item) %>%
  nest() %>%
  ungroup() %>%
  mutate(
    emo_item = ifelse(emo_item == "Often felt sad.", "Sad", emo_item)
    , data = map(data, rescale_time)
  ) %>%
  filter(map_dbl(data, nrow) >= 35) %>%
  rename(SID = SID2)

save_dat_fun <- function(x, emo, sid, sample){
  if(sd(x$emo_value) > 1){
  write_csv(x, file = sprintf("../02-data/02-clean-data/rq1-robustness/csv/%s-%s-%s.csv", sample, sid, emo))
  save(x, file = sprintf("../02-data/02-clean-data/rq1-robustness/rdata/%s-%s-%s.RData", sample, sid, emo))
  return("variable")
  } else {
    return("no variability")
  }
}

nested_rq1 <- nested_rq1 %>%
  mutate(dat = pmap_chr(list(data, emo_item, SID, sample), save_dat_fun))
save(nested_rq1, file = "../02-data/02-clean-data/rq1/nested_rq1.RData")

low_var_rq1 <- nested_rq1 %>% filter(dat == "no variability") %>% select(-data)
```

#### RQ2

```{r}
rescale_time <- function(x){
  x %>%
    arrange(Date) %>%
    mutate(time = difftime(Date, Date[1], unit = "hours")) %>%
    as.data.frame()
}

nested_dat <- emo_long %>%
  rename(emo_item = item, emo_value = value) %>%
  full_join(
    ergoals_long %>%
      rename(ergoal = item_name, ergoal_value = value) %>%
      select(-item, -category)
    ) %>%
  drop_na(emo_item, emo_value, ergoal, ergoal_value, Date) %>%
  # filter(er_use == 1) %>%
  mutate(
    time = difftime(Date, Date[1], unit = "hours")
    , ExER = emo_value*ergoal_value 
    , SID2 = SID
    ) %>%
  group_by(SID2, sample, ergoal, emo_item) %>%
  nest() %>%
  ungroup() %>%
  mutate(
    emo_item = ifelse(emo_item == "Often felt sad.", "Sad", emo_item)
    , data = map(data, rescale_time)
  ) %>%
  filter(map_dbl(data, nrow) >= 35) %>%
  rename(SID = SID2)

save_dat_fun <- function(x, emo, goal, sid, sample){
  if(sd(x$emo_value) > 1 & sd(x$ergoal_value) > 1){
    write_csv(x, file = sprintf("../02-data/02-clean-data/rq2/csv/%s-%s-%s-%s.csv", sample, sid, emo, goal))
    save(x, file = sprintf("../02-data/02-clean-data/rq2/rdata/%s-%s-%s-%s.RData", sample, sid, emo, goal))
    return("variable")
  } else {
    return("no variability")
  }
}

nested_dat <- nested_dat %>%
  mutate(dat = pmap_chr(list(data, emo_item, ergoal, SID, sample), save_dat_fun))
save(nested_dat, file = "../02-data/02-clean-data/rq2-robustness/nested_dat.RData")

low_var_rq2 <- nested_dat %>% filter(dat == "no variability") %>% select(-data)
```

### Reliability

```{r}
ml_rel_fun <- function(x){
  y <- x %>%
    distinct() %>%
    pivot_wider(
      names_from = "item_name"
      , values_from = "value"
    ) %>% drop_na() %>% data.frame()
  multilevelTools::omegaSEM(
    data = y
    , id = "SID"
    , items = colnames(y)[-c(1:2)]
    , savemodel = TRUE
  )$Results
}

ema_emo_rel <- ema_long %>%
  filter(category == "emotion") %>%
  select(
    SID, Date, item, item_name, value
    ) %>%
  group_by(item) %>%
  nest() %>%
  ungroup() %>%
  mutate(rel = map(data, ml_rel_fun)) %>% 
  select(-data) %>%
  unnest(rel) %>%
  # mutate(est = sprintf("%.2f<br>[%.2f, %.2f]", est, ci.lower, ci.upper)) %>%
  select(-ci.lower, -ci.upper) %>%
  pivot_wider(names_from = "label", values_from = "est")

```

### Descriptives

```{r}
var_labs <- cb %>% 
  filter(category %in% c("emotion", "emotion regulation goals")) %>%
  select(category, item_name, long_name)

ema_desc <- emo_long %>%
  select(SID, Date, sample, item, value) %>%
  mutate(category = "Emotion") %>%
  full_join(
    ergoals_long %>%
      select(SID, Date, sample, item, value) %>%
      mutate(category = "Emotion Regulation Goals")
  ) %>%
  group_by(SID, sample, item) %>%
  filter(n() >= 35) %>%
  ungroup() %>%
  group_by(SID, item) %>%
  summarize(
    mean = mean(value, na.rm = T)
    , sd = sd(value, na.rm = T)
  ) %>%
  group_by(item) %>%
  summarize(
    mean_mean = mean(mean)
    , sd_mean = sd(mean)
    , min_sd = min(sd)
    , max_sd = max(sd)
  ) %>%
  left_join(ema_emo_rel) %>%
  mutate(item = factor(item, levels = var_labs$item_name, labels = var_labs$long_name)) %>%
  arrange(item) %>%
  mutate(mean = sprintf("%.2f<br>(%.2f)", mean_mean, sd_mean)) %>%
  select(item, mean, omega_within, omega_between)

varL <- tibble(vars = unique(ema_desc$item)) %>% mutate(n = 1:n(), num = paste(sprintf("%i. %s", n, vars)))

```

### Zero-Order Correlations

```{r}
# make N = 1 correlation tables
cor_tab_fun <- function(r, SID, sample){
  print(SID)
  r$r <- apply(r$r, c(1,2), function(x) sprintf("%.2f", x))
  r$r[which(r$p < .05, arr.ind = T)] <- 
    sprintf("<strong>%s</strong>" ,r$r[which(r$p < .05, arr.ind = T)])
  r$r[upper.tri(r$r, diag = F)] <- ""
  diag(r$r) <- "--"
  r <- r$r %>% data.frame %>%
    rownames_to_column("item") %>%
    as_tibble() %>%
    mutate_all(as.character) %>%
    select(item, everything()) %>%
    mutate(item = factor(item, levels = var_labs$item_name, labels = var_labs$long_name),
           item = factor(item, levels = varL$vars, labels = varL$num)) %>%
    arrange(item) %>%
    select(item, one_of(var_labs$item_name)) %>%
    setNames(c("item", 1:11)) %>%
    kable(
    .
    , "html"
    , escape = F
    , col.names = c("EMA Variable", paste0(1:11, "."))
    , align = c("l", rep("c", 11))
    , caption = sprintf("<strong>Table SX</strong><br><em>Within-Person Correlations for Participant %s in %s for all EMA variables</em>", SID, sample)
  ) %>%
  kable_classic(html_font = "Times")
  save_kable(r, file = sprintf("../05-results/tables/px-r-tabs/%s-%s.html", sample, SID))
  return(r)
}

# estimate and reshape correlations for between-person summaries
cor_fun <- function(x){
  r <- cor(x, use = "pairwise")
  r[upper.tri(r, diag = F)] <- NA
  fisherz(r) %>%
    data.frame() %>%
    rownames_to_column("V1") %>%
    pivot_longer(
      cols = -V1
      , names_to = "V2"
      , values_to = "z"
      , values_drop_na = T
      )
}

sc_ema_r <- emo_long %>%
  select(SID, Date, sample, item, value) %>%
  full_join(
    ergoals_long %>%
      select(SID, Date, sample, item, value)
  ) %>%
  group_by(SID, sample, item) %>%
  filter(n() >= 35) %>%
  pivot_wider(
    names_from = "item"
    , values_from = "value"
      ) %>%
  select(-Date) %>%
  group_by(sample, SID) %>%
  nest() %>%
  ungroup() %>%
  mutate(r_tab = map(data, ~corr.test(., use = "pairwise"))
         , r_tab = pmap(list(r_tab, SID, sample), cor_tab_fun)
         , r = map(data, cor_fun))

## reshape summaries for manuscript table
sc_ema_r <- sc_ema_r %>%
  select(-data, -r_tab) %>%
  unnest(r) %>%
  mutate(z = ifelse(is.infinite(z), NA, z)) %>%
  group_by(V1, V2) %>%
  summarize(
    sd = sd(z, na.rm = T)
    , z = mean(z, na.rm = T)
    , z = fisherz2r(z)
    , sd = fisherz2r(sd)
    ) %>%
  mutate(
    z = ifelse(is.nan(z) | is.na(z), "--", ifelse(abs(z) < .01, sprintf("%.3f", z), sprintf("%.2f", z)))
    , sd = ifelse(is.nan(sd) | is.na(sd), "", sprintf("%.2f", sd))
    , z = ifelse(V1 != V2 & !(V1 == "Often felt sad." & V2 == "Often.felt.sad."), sprintf("%s<br>(%s)", z, sd), "--")
    ) %>%
  select(-sd) %>%
  ungroup() %>%
  pivot_wider(
    names_from = "V2"
    , values_from = "z"
  ) %>%
  mutate(V1 = factor(V1, levels = var_labs$item_name, labels = var_labs$long_name),
         V1 = factor(V1, levels = varL$vars, labels = varL$num)) %>%
  arrange(V1) %>%
  select(V1, Negative, Positive, Angry, Afraid, Happy, Excited, Proud, Guilty, Content, `Often.felt.sad.`, goalPos, goalNeg) %>%
  setNames(c("variable", 1:12)) %>%
  left_join(
    ema_desc %>% mutate(item = factor(item, levels = varL$vars, labels = varL$num)) %>%
      rename(variable = item)
    ) %>%
  select(variable, mean, omega_within, omega_between, everything())
```

## Baseline Data

```{r}
bl_df <- read_csv("../02-data/01-raw-data/erdyn-baseline-2025-06-09.csv") %>%
  mutate(StartDate = parse_date_time(StartDate, orders = c("mdy_hM", "ymd_HMS")))

bl_items <- cb %>% filter(level == "Baseline" & category != "proc") 

reverse <- function(x, maxi){
  y <- (maxi+1) - x
  return(y)
}

bl_long <- bl_df %>%
  pivot_longer(
    cols = c(starts_with("CESD"), starts_with("SWL"))
    , names_to = "old_name"
    , values_to = "value"
  ) %>%
  left_join(
    bl_items %>%
      select(item, old_name = `Item #`, reverse)
  ) %>%
  mutate(maxi = ifelse(item == "CESD", 4, 7) 
         , value = ifelse(!is.na(reverse) & reverse == -1, reverse(value, maxi), value)) %>%
  select(-reverse)
```

### Reliability

```{r}
omega_fun <- function(d){
  x <- d %>% 
    pivot_wider(
      names_from = "old_name"
      , values_from = "value"
      ) %>%
    select(-SID, -StartDate, -sample) %>%
    as.matrix()
  
  om <- omega(x)
  return(tibble(alpha = om$alpha, omega = om$omega.tot))
}

# relative to the alpha function, those alphas don't look right, so probably ignore? 

nested_meas <- bl_long %>%
  select(SID, StartDate, sample, item, old_name, value) %>%
  group_by(item) %>%
  nest() %>%
  ungroup() %>%
  mutate(rel = map(data, omega_fun)) %>%
  unnest(rel) %>%
  mutate(omega = sprintf("%.2f", omega))
```

```{r}
wb_long <- bl_long %>%
  group_by(SID, StartDate, sample, item, maxi) %>%
  summarize(value = mean(value, na.rm = T)) %>%
  ungroup() %>%
  rename(wb_item = item, wb_value = value) %>%
  mutate(wb_value = pomp(wb_value, mini = 1, maxi = maxi))
```

### Demographics

```{r}
all_subs <- ema_df %>%
  group_by(SID, sample) %>%
  filter(n() >= 1) %>%
  ungroup() %>%
  select(SID, sample) %>%
  distinct() 

ema_df %>%
  group_by(SID, sample) %>%
  filter(n() >= 1) %>%
  group_by(sample) %>%
  tally()

subs

# race 
bl_df %>%
  select(SID:YOB, sample) %>%
  filter(SID %in% all_subs$SID) %>%
  # filter(SID %in% subs) %>%
  mutate(
    race = map_chr(strsplit(as.character(race), ""), ~paste(., collapse = ","))
    , race_mixed = ifelse(grepl("[,]", race) | (race != 1 & !is.na(ethnicity)), 1, 0)
    , race_text = mapvalues(race, c(1:7), c("White", "Black", "Other", "East Asian",  "Other", "Other", "Other"))
    , race_text = ifelse(grepl("[,]", race_text), "Mixed Race / Ethnicity", race_text)
    , race_text = ifelse(!is.na(ethnicity) & race == 1, "Hispanic", race_text)
    ) %>%
  filter(sample == "pp") %>%
  group_by(race_text) %>% 
  tally() %>%
  ungroup() %>%
  mutate(perc = n/sum(n)*100) %>%
  select(-n)

# age
bl_df %>%
  mutate(age= year(StartDate)-as.numeric(YOB)) %>%
  select(SID, sample, age) %>%
  filter(SID %in% all_subs$SID) %>%
  # filter(SID %in% subs) %>%
  group_by(sample) %>%
  summarize(
    n = n()
    , mean = mean(age, na.rm = T)
    , sd = sd(age, na.rm = T)
    )

# gender 
bl_df %>%
  select(SID:YOB, sample) %>%
  mutate(gender = mapvalues(gender, c(1, 2, 5, 8, 9), c("Man", "Woman", "Non-binary", "Other", NA))) %>%
  # filter(SID %in% all_subs$SID) %>%
  filter(SID %in% subs) %>%
  filter(sample == "pp") %>%
  group_by(gender) %>% 
  tally() %>%
  ungroup() %>%
  mutate(perc = n/sum(n)*100) %>%
  select(-n)

bl_df %>%
  filter(sample == "cpe") %>%
  # filter(SID %in% all_subs$SID) %>%
  filter(SID %in% subs) %>%
  mutate(age = 2024-as.numeric(YOB)) %>%
  pull(age) %>%
  describe()
```

```{r}
subs <- nested_dat %>% select(SID) %>% distinct() %>% pull()

var_labs <- 
  tibble(category = "demo", item_name = "age", long_name = "Age") %>%
  full_join(
    cb %>% 
      filter(category %in% c("emotion", "emotion regulation goals")) %>%
      select(category, item_name, long_name) %>%
      filter(item_name %in% c("Negative", "Positive", "goalNeg", "goalPos")) %>% 
      arrange(item_name)
  ) %>%
  full_join(
    cb %>% 
      filter(category %in% c("Correlates")) %>%
      select(category, item_name = item, long_name) %>%
      distinct()
  )

wb_long %>%
  select(SID, sample, wb_item, wb_value) %>%
  pivot_wider(
    names_from = "wb_item"
    , values_from = "wb_value"
    , values_fn = mean
      ) %>%
  inner_join(
    bl_df %>%
      mutate(age = year(StartDate)-as.numeric(YOB)) %>%
      select(SID, sample, age)
  )

ema_pm <- emo_long %>%
  select(SID, Date, sample, item, value) %>%
  mutate(category = "Emotion") %>%
  full_join(
    ergoals_long %>%
      select(SID, Date, sample, item, value) %>%
      mutate(category = "Emotion Regulation Goals")
  ) %>%
  filter(item %in% c("Negative", "Positive", "goalNeg", "goalPos")) %>%
  group_by(SID, sample, item) %>%
  filter(n() >= 35) %>%
  ungroup() %>%
  group_by(SID, sample, item) %>%
  summarize(
    mean = mean(value, na.rm = T)
  ) %>%
  ungroup() 

ema_desc <- ema_pm %>%
  full_join(
    wb_long %>%
      select(SID, sample, item = wb_item, mean = wb_value)
  ) %>%
  full_join(
    bl_df %>%
      mutate(item = "age", mean = year(StartDate)-as.numeric(YOB)) %>%
      select(SID, sample, item, mean)
  ) %>%
  filter(SID %in% subs) 

demo_tab <- ema_desc %>%
  group_by(item) %>%
  summarize(
    sd = sd(mean, na.rm = T)
    , mean = mean(mean, na.rm = T)
  ) %>%
  left_join(
    ema_emo_rel %>%
    mutate_at(vars(matches("omega")), ~sprintf("%.2f", .)) %>%
    full_join(nested_meas %>% select(item, omega_between = omega))
    ) %>%
  mutate(variable = factor(item, levels = var_labs$item_name, labels = var_labs$long_name)) %>%
  arrange(variable) %>%
  mutate(msd = sprintf("%.2f (%.2f)", mean, sd)) %>%
  select(variable, msd, omega_between, omega_within)

bl_r <- ema_desc %>%
  pivot_wider(
    names_from = "item"
    , values_from = "mean"
    , values_fn = mean
    ) %>%
  select(age, Negative, Positive, goalNeg, goalPos, SWL, DEP) %>%
  corr.test(., use = "pairwise")


cor_tab_fun <- function(r){
  r$r <- apply(r$r, c(1,2), function(x) sprintf("%.2f", x))
  r$r[which(r$p < .05, arr.ind = T)] <- 
    sprintf("<strong>%s</strong>" ,r$r[which(r$p < .05, arr.ind = T)])
  r$r[upper.tri(r$r, diag = F)] <- ""
  diag(r$r) <- "--"
  r <- r$r %>% data.frame %>%
    rownames_to_column("variable") %>%
    as_tibble() %>%
    mutate_all(as.character) %>%
    select(variable, everything()) 
  return(r)
}

bl_r_tab <- cor_tab_fun(bl_r) %>%
  mutate(variable = factor(variable, levels = var_labs$item_name, labels = var_labs$long_name)) %>%
  arrange(variable) %>%
  setNames(c("variable", 1:7)) %>%
  left_join(demo_tab ) %>%
  select(variable, msd, omega_within, omega_between, everything()) %>%
  mutate_at(vars(omega_between, omega_within), ~ifelse(is.na(.), "--", .)) %>%
  mutate(variable = sprintf("%i. %s", 1:n(), variable)) %>%
  kable( 
    .
    , "html"
    , escape = F
    , col.names = c("Variable", "M (SD)", "$\\omega_w$ [CI]", "$\\omega_b$ [CI]", paste0(1:7, "."))
    , align = c("l", rep("c", 9))
    , caption = "<strong>Table X</strong><br><em>Between-Person Descriptive Statistics and Correlations</em>"
  ) %>%
  kable_classic(html_font = "Times") %>%
  add_header_above(c(" " = 1, "Descriptives" = 3, "Correlations" = 7))
bl_r_tab
save_kable(bl_r_tab, file = "../05-results/tables/tab-x-bp-desc.html")
```
