==> 04-scripts/03-summaries-new.qmd <==
---
title: "ctsem Models"
format: html
editor: visual
editor_options: 
  chunk_output_type: console
---

# Load in Model Summaries

```{r}
loadRData <- function(file, obj, rq){
  file <- sprintf("../05-results/%s/fx/%s", rq, file)
  load(file)
  get(ls()[grepl(obj, ls())])
}

nested_sum <- tibble(model = c("RQ1", "RQ2-int", "RQ2-main")) %>%
  mutate(file = map(model, ~list.files(sprintf("../05-results/%s/fx", .)))) %>%
  unnest(file) %>%
  mutate(
    fx = pmap(list(file, "fx", model), loadRData)
    , fx = map(fx, unnest)
    ) %>%
  separate(file, c("study", "SID", "emo_item", "ergoal", "er_exc"), sep = "-") %>%
  mutate_at(vars(emo_item, ergoal, er_exc), ~str_remove_all(., ".RData")) %>%
  mutate(
    er_exc = ifelse(is.na(er_exc), ergoal, er_exc)
    , ergoal = ifelse(ergoal %in% c("all", "er"), NA_real_, ergoal)
    )

# exclude by person
# some sort of group_by and filter(any())
summ_df <- nested_sum %>% 
  unnest(fx) %>% 
  group_by(model, study, SID, emo_item, ergoal, er_exc) %>%
  mutate(flag = ifelse(
    (model == "RQ1" & grepl("beta_E", parameter) & sign(mean) == 1) |
    (grepl("mu", parameter) & (mean > 10 | mean < 0)) |
    sd == 0 | 
    sd > 10
    , "implausible", "plausible"))

summ_df <- summ_df %>% 
  filter(all(flag == "plausible")) %>%
  ungroup() %>%
  mutate(exclusion = "person") %>%
  full_join(
    summ_df %>%
      filter(flag == "plausible") %>%
      mutate(exclusion = "parameter")
  )

## counts across exclusions and emotions
summ_df %>% 
  group_by(model, emo_item, ergoal, SID, er_exc, exclusion) %>%
  slice_head(n = 1) %>%
  group_by(model, emo_item, ergoal, er_exc, exclusion) %>%
  tally() %>% 
  pivot_wider(
    names_from = "emo_item"
    , values_from = "n"
  )

summ_df %>%
  mutate(sig = ifelse(sign(lower) == sign(upper), "sig", "ns")) %>%
  group_by(model, emo_item, ergoal, er_exc, exclusion, parameter, sig) %>%
  tally() %>% 
  pivot_wider(
    names_from = "emo_item"
    , values_from = "n"
  ) %>%
  filter(grepl("strength", parameter))

```

# Meta-Analytic Summaries

## Run Models

```{r}
ma_fun <- function(d, mod, emo, er, param, exc, er_exc){
  m <- rma(
    yi = mean
    , sei = sd
    , data = d
    , control=list(stepadj=0.5, maxiter=1000)
    )
  save(m, file = sprintf("../05-results/%s/ma-models/%s-%s-%s-%s-%s.RData", mod, emo, er, param, exc, er_exc))
  
  fx <- tidy(m, conf.int = T) %>%
    full_join(
      tribble(
        ~term, ~estimate,
        "sd", sqrt(m$tau2),
        "Q", m$QE,
        "p", m$QEp
        )
      )
  
  save(fx, file = sprintf("../05-results/%s/ma-fx/%s-%s-%s-%s-%s.RData", mod, emo, er, param, exc, er_exc))
  return(NULL)
}

plan(multisession(workers=12L))
ma_nested <- summ_df %>%
  mutate(mean = ifelse(grepl("beta_E|strength", parameter), abs(mean), mean)) %>%
  group_by(model, emo_item, ergoal, parameter, exclusion, er_exc) %>% 
  filter(!parameter %in% c("mu_E", "sigma_E", "sigma_p")) %>%
  nest() %>%
  ungroup() %>%
  arrange(model, emo_item, ergoal, parameter) %>%
  mutate(ma = 
    future_pmap(
      list(data, model, emo_item, ergoal, parameter, exclusion, er_exc)
      # , .f = possibly(ma_fun, NA_real_)
      , .f = ma_fun
      , .progress = T
      , .options = furrr_options(packages = c("brms", "broom.mixed"))
    ))
closeAllConnections()
```

## Load in Results

```{r}
loadRData <- function(file, obj, rq){
  file <- sprintf("../05-results/%s/ma-fx/%s", rq, file)
  load(file)
  get(ls()[grepl(obj, ls())])
}

ma_nested <- tibble(model = c("RQ1", "RQ2-int", "RQ2-main")) %>%
  mutate(file = map(model, ~list.files(sprintf("../05-results/%s/ma-fx", .)))) %>%
  unnest(file) %>%
  mutate(fx = pmap(list(file, "fx", model), loadRData)) %>%
  separate(file, c("emo_item", "ergoal", "parameter", "exclusion", "er_exc"), sep = "-") %>%
    mutate(
      er_exc = str_remove_all(er_exc, ".RData")
      , ergoal = ifelse(ergoal == "NA", NA_real_, ergoal)
      ) %>%
  # inner_join(ma_nested %>% select(model:er_exc)) %>%
  distinct() %>%
  left_join(
    summ_df %>%
      mutate(mean = ifelse(grepl("beta_E|strength", parameter), abs(mean), mean)) %>%
      group_by(model, emo_item, ergoal, parameter, exclusion, er_exc) %>% 
      filter(!parameter %in% c("mu_E", "sigma_E", "sigma_p")) %>%
      nest() %>%
      ungroup() %>%
      arrange(model, emo_item, ergoal, parameter) 
  )
```

## Reshape Results

```{r}
nested_ma <- ma_nested %>%
  select(-data) %>% 
  unnest(fx) %>%
  filter(term != "sd__Observation") %>%
  mutate(term = mapvalues(term, c("overall", "sd"), c("M", "SD"))) %>%
  select(-type, -statistic) %>%
  mutate(
    sig = ifelse(sign(conf.low) == sign(conf.high), "sig", "ns")
    , est = ifelse(term == "M", sprintf("%.2f<br>[%.2f,%.2f]", estimate, conf.low, conf.high)
            , ifelse(term == "Q", sprintf("%.2e", estimate)
            , ifelse(term == "p", ifelse(estimate < .001, "&lt;.001", sprintf("%.2f", estimate))
                   , sprintf("%.2f", estimate))))
    ) %>%
  select(-(estimate:sig)) %>%
  pivot_wider(
    names_from = "term"
    , values_from = "est"
  ) %>%
  inner_join(
    ma_nested %>%
      select(-fx) %>%
      unnest(data) %>% 
      group_by(model, emo_item, ergoal, parameter, exclusion, er_exc) %>%
      summarize_at(vars(mean), lst(min, max)) %>%
      ungroup() %>%
      mutate_at(vars(min, max), ~sprintf("%.2f", .)) 
  ) %>%
  filter(!grepl("sigma", parameter)) %>%
  mutate(
    focal = ifelse(emo_item %in% c("Negative", "Positive"), "focal", "sup")
    , parameter = mapvalues(
      parameter, c("beta_E", "location"), 
      c("Attractor Strength", "Attractor Location")
      )
    ) %>%
  filter(!(model == "RQ2-int" & (grepl("beta", parameter) | !grepl("_", parameter)))) %>%
  group_by(model, focal, exclusion, er_exc) %>%
  nest() %>%
  ungroup()
```

## Tables

### RQ1 & RQ2 Main Effects

```{r}
ma_tab_fun <- function(d, rq, focal, exc, er_exc){
  if(rq == "RQ1"){
    d <- d %>% select(-ergoal)
    cols <- c("Emotion", "Parameter", "M [CI]", "SD [CI]", "Q", "p", "Min", "Max")
    al <- c("l", "l", rep("c", 6))
    cap <- "Simple Emotion Systems (RQ1)"
  } else {
    cols <- c("Emotion", "ER Goal", "Parameter", "M [CI]", "SD [CI]", "Q", "p", "Min", "Max")
    al <- c("l", "l", "l", rep("c", 6))
    cap <- "Main Effects of Emotion Regulation Goals (RQ2 main)"
  }
  cap <- sprintf("<strong>Table SX</strong><br><em>Meta-Analytic Estimates and Descriptive Statistics of %s </em>", cap)
  tab <- d %>%
    kable(
      .
      , round = 2
      , "html"
      , col.names = cols
      , escape = F
      , align = al
      , cap = cap
    ) %>%
    kable_classic(html_font = "Times", full_width = F) %>%
    collapse_rows(1)
  save_kable(tab, file = sprintf("../05-results/tables/%s/key-terms/tabSx-%s-%s-%s-%s-ma.html", rq, focal, str_to_lower(rq), exc, er_exc))
  return(tab)
}

nested_ma_main <- nested_ma %>%
  filter(model != "RQ2-int") %>%
  mutate(tab = pmap(list(data, model, focal, exclusion, er_exc), ma_tab_fun))
nested_ma_main %>% filter(focal == "focal" & exclusion == "parameter" & model == "RQ1") %>% pull(tab)
```

### RQ2 Interactions

```{r}
ma_tab_fun <- function(d, rq, focal, exclusion, er_exc){
  cap <- "Interaction between Emotion and Emotion Regulation Goals (RQ2 interaction)"
  cap <- sprintf("<strong>Table SX</strong><br><em>Meta-Analytic Estimates and Descriptive Statistics of %s </em>", cap)
  d <- d %>% 
    mutate(
      ergoal = mapvalues(ergoal, c("goalNeg", "goalPos"), c("Goal to Change Negative Emotions", "Goal to Change Positive Emotions"))
      , emo_item = paste(emo_item, "Emotion")
      ) %>%
    separate(parameter, c("param", "ergoal_lev")) %>%
    mutate_at(vars(param, ergoal_lev), str_to_title) %>% 
    mutate(ergoal_lev = factor(ergoal_lev, c("Decrease", "Maintain", "Increase"))) %>% 
    arrange(emo_item, ergoal, param, ergoal_lev) %>%
    unite(comb, emo_item, ergoal, sep = ": ")
  
  cols <- c("Parameter","Emotion Regulation Goal", "M [CI]", "SD [CI]", "Q", "p", "Min", "Max")
  
  rs <- d %>% group_by(comb) %>% tally() %>% 
      mutate(end = cumsum(n), start = lag(end) + 1, start = ifelse(is.na(start), 1, start))
  
  tab <- d %>%
    select(-comb) %>% 
    # mutate_at(vars(Q, p), ~ifelse(ergoal_lev == "Decrease", ., "")) %>%
    kable(
      .
      , round = 2
      , "html"
      , col.names = cols
      , escape = F
      , align = c(rep("l", 2), rep("c", 6))
      , cap = cap
    ) %>%
    kable_classic(html_font = "Times", full_width = F) #%>%
    # collapse_rows(1)
  for (i in 1:nrow(rs)){
    tab <- tab %>% 
      kableExtra::group_rows(rs$comb[i], rs$start[i], rs$end[i]) 
  }
  tab
  save_kable(tab, file = sprintf("../05-results/tables/RQ2-int/key-terms/tabSx-rq2int-%s-%s-%s-ma.html", focal, exclusion, er_exc))
  return(tab)
}

nested_ma_int <- nested_ma %>%
  filter(model == "RQ2-int") %>%
  mutate(tab = pmap(list(data, model, focal, exclusion, er_exc), ma_tab_fun))

nested_ma_int %>% filter(focal == "focal" & exclusion == "parameter" & er_exc == "all") %>% pull(tab)
```

# RQ3a Correlations
```{r}
cor_tab_fun <- function(r){
  r$r <- apply(r$r, c(1,2), function(x) sprintf("%.2f", x))
  r$r[which(r$p < .05, arr.ind = T)] <- 
    sprintf("<strong>%s</strong>" ,r$r[which(r$p < .05, arr.ind = T)])
  r$r[upper.tri(r$r, diag = F)] <- ""
  diag(r$r) <- "--"
  r <- r$r %>% data.frame %>%
    rownames_to_column("item") %>%
    as_tibble() %>%
    mutate_all(as.character) %>%
    select(item, everything()) %>%
    mutate(item = sprintf("%i. %s", seq(1, ncol(r$r)), item)) %>%
    setNames(c("item", seq(1, ncol(r$r)))) 
  return(r)
}
```

## RQ3a - Correlations among parameters

```{r}
nested_r <- summ_df %>% 
  filter(model %in% c("RQ1", "RQ2-int") & grepl("strength|location", parameter)) %>%
  # filter((parameter == "location" & mean > 0) | parameter == "beta_E") %>%
  select(model, SID, emo_item, ergoal, parameter, mean, exclusion, er_exc) %>%
  # mutate(parameter = mapvalues(parameter, c("beta_E"), c("strength"))) %>%
  mutate(focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")) %>%
  group_by(model, focal, exclusion, er_exc) %>%
  nest() %>%
  ungroup()
  # mutate(data = map(data, ~(. %>% 
  #     pivot_wider(
  #       names_from = c("emo_item", "parameter")
  #       , values_from = "mean"
  #     )



cor_tab_html_fun <- function(tab, focal, rq, exc, er_exc){
  len <- nrow(tab)
  tab <- tab %>%
    separate(item, c("num", "emo_item", "param"), sep = "[ ]|_") %>%
    kable(
      .
      , "html"
      , escape = F
      , col.names = c(" ", "Variable", "Parameter", paste0(1:len, "."))
      , align = c("l", "l", "l", rep("c", len))
      , caption = sprintf("<strong>Table SX</strong><br><em>RQ3a: Correlations Among %s Parameters</em>", rq)
    ) %>%
    kable_classic(html_font = "Times")
  save_kable(tab, file = sprintf("../05-results/tables/%s/correlations/%s-%s-%s.html", rq, focal, exc, er_exc))
  return(tab)
}

lm_fun <- function(d) tidy(effectsize::standardize(lm(strength ~ location + location2, data = d)), conf.int = T)

quad_fun <- function(d){
  d %>% 
    pivot_wider(
      names_from = "parameter"
      , values_from = "mean"
    ) %>%
    mutate(location2 = location^2) %>%
    group_by(emo_item) %>%
    nest() %>%
    ungroup() %>%
    mutate(m = map(data, lm_fun)) %>%
    select(-data) %>% 
    unnest(m) %>%
    filter(term == "location2")
}

nested_rq1_r <- summ_df %>% 
  filter(model == "RQ1" & parameter %in% c("beta_E", "location")) %>%
  filter((parameter == "location" & mean > 0) | parameter == "beta_E") %>%
  select(SID, exclusion, er_exc, emo_item, parameter, mean) %>%
  mutate(
    focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")
    , parameter = mapvalues(parameter, c("beta_E"), c("strength"))
    ) %>%
  arrange(focal, emo_item, parameter) %>%
  group_by(focal, exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  mutate(
    r2 = map(data, quad_fun)
    , data = map(data, ~(.) %>% 
      pivot_wider(
        names_from = c("emo_item", "parameter")
        , values_from = "mean"
      ))
    , r = map(data, ~(.) %>% select(-SID) %>% corr.test())
    , rtab = map(r, cor_tab_fun)
    , rtab_html = pmap(list(rtab, focal, "RQ1", exclusion, er_exc), cor_tab_html_fun)
  )

nested_rq1_r %>% filter(focal == "focal" & exclusion == "parameter" & er_exc == "all") %>% pull(rtab_html)
```

```{r}
col_ord <- cols <- c(
  "Negative_goalNeg_location_decrease", "Negative_goalNeg_location_maintain"
  , "Negative_goalNeg_location_increase", "Negative_goalNeg_strength_decrease"
  , "Negative_goalNeg_strength_maintain", "Negative_goalNeg_strength_increase"
  , "Positive_goalPos_location_decrease", "Positive_goalPos_location_maintain"
  , "Positive_goalPos_location_increase", "Positive_goalPos_strength_decrease"
  , "Positive_goalPos_strength_maintain", "Positive_goalPos_strength_increase"
  )

cor_tab_html_fun <- function(tab, focal, rq, exclusion, er_exc){
  tab <- tab %>%
    separate(
      item
      , c("num", "emo_item", "ergoal", "parameter", "ergoal_lev")
      , sep = "[ ]|_")
  len <- nrow(tab)
  tab <- tab %>%
    select(-emo_item, -ergoal) %>%
    kable(
      .
      , "html"
      , escape = F
      , col.names = c(" ", "Variable", " ", paste0(1:len, "."))
      , align = c("l", rep("c", len + 2))
      , caption = sprintf("<strong>Table SX</strong><br><em>RQ3a: Correlations Among %s Parameters</em>", rq)
    ) %>%
    kable_classic(html_font = "Times") %>%
    kableExtra::group_rows("Negative Emotions: Change Negative Emotions", 1, len/2) %>%
    kableExtra::group_rows("Positive Emotions: Change Positive Emotions", len/2+1, len) %>%
    add_header_above(c(" " = 3, "Negative Emotions: Change Negative Emotions" = len/2, "Positive Emotions: Change Positive Emotions" = len/2)) %>%
    add_header_above(c(" " = 3, "Zero-Order Correlations" = len))
  save_kable(tab, file = sprintf("../05-results/tables/%s/correlations/%s-%s-%s.html", rq, focal, exclusion, er_exc))
  return(tab)
}

nested_rq2int_r <- summ_df %>% 
  filter(model == "RQ2-int" & grepl("location_|strength_", parameter)) %>%
  select(SID, exclusion, er_exc, emo_item, ergoal, parameter, mean) %>%
  filter((ergoal == "goalNeg" & emo_item %in% c("Negative")) | 
         (ergoal == "goalPos" & emo_item %in% c("Positive"))) %>%
  mutate(
    focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")
    ) %>%
  arrange(focal, emo_item, ergoal, parameter) %>%
  group_by(focal, exclusion, er_exc) %>%
  nest() %>%
  mutate(
    data = map(data, ~(.) %>% 
      pivot_wider(
        names_from = c("emo_item", "ergoal", "parameter")
        , values_from = "mean"
      ))
    , r = map(data, ~(.) %>% select(one_of(col_ord)) %>% corr.test())
    , rtab = map(r, cor_tab_fun)
    , rtab_html = pmap(list(rtab, focal, "RQ2-int", exclusion, er_exc), cor_tab_html_fun)
  )

# nested_rq2int_r <- summ_df %>% 
#   filter(model == "RQ2-int" & grepl("location_|strength_", parameter)) %>%
#   select(SID, exclusion, emo_item, ergoal, parameter, mean) %>%
#   filter((ergoal == "goalNeg" & emo_item %in% c("Negative", neg_emo_items$long_name)) | 
#          (ergoal == "goalPos" & emo_item %in% c("Positive", pos_emo_items$long_name))) %>%
#   mutate(
#     focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")
#     ) %>%
#   arrange(focal, emo_item, ergoal, parameter) %>%
#   group_by(focal, exclusion) %>%
#   nest() %>%
#   mutate(
#     data = map(data, ~(.) %>% 
#       pivot_wider(
#         names_from = c("emo_item", "ergoal", "parameter")
#         , values_from = "mean"
#       ))
#     , r = map(data, ~(.) %>% select(-SID) %>% corr.test())
#     , rtab = map(r, cor_tab_fun)
#     , rtab_html = pmap(list(rtab, focal, "RQ2-int"), cor_tab_html_fun)
#   )

(nested_rq2int_r %>% filter(exclusion == "parameter" & er_exc == "all"))$rtab_html
```

# RQ3b Meta-Regressions

## Run Models

```{r}
mr_fun <- function(d, mod, emo, er, param, type, o_item, exc, er_exc){
  frm <- if(type == "linear") "~ o_value" else "~ o_value + o_value_2"
  m <- rma(
    yi = mean
    , sei = sd
    , mods = as.formula(frm)
    , data = d
    , control=list(stepadj=0.25, maxiter=5000)
  )
  save(m, file = sprintf("../05-results/%s/mr-models/%s-%s-%s-%s-%s-%s-%s.RData", mod, emo, er, param, o_item, type, exc, er_exc))
  
  fx <- tidy(m, conf.int = T)
  save(fx, file = sprintf("../05-results/%s/mr-fx/%s-%s-%s-%s-%s-%s-%s.RData", mod, emo, er, param, o_item, type, exc, er_exc))
  return(NULL)
}

z_score <- function(x) (x - mean(x, na.rm = T))/sd(x, na.rm = T)

mr_nested <- crossing(
  type = c("linear", "quadratic")
  , ma_nested# %>% select(-fx)
) %>%
  crossing(
    .
    , wb_long %>%
      rename(o_value = wb_value, o_item = wb_item) %>%
      mutate(cat = "wb") %>%
      full_join(
        erstrat %>% 
          select(SID, er_strat_rep:avg_talkPos) %>%
          pivot_longer(
            cols = -SID
            , names_to = "o_item"
            , values_to = "o_value"
            ) %>%
          mutate(cat = "ER")
        ) %>%
      group_by(o_item) %>%
      mutate(
        o_value = z_score(o_value)
        , o_value_2 = o_value^2
        , SID = as.character(SID)
        ) %>%
      group_by(cat, o_item) %>%
      nest() %>%
      ungroup() %>%
      rename(cor_data = data)
  ) %>%
  filter(!(cat == "ER" & type == "quadratic")) %>%
  mutate(data = map2(data, cor_data, ~(.x) %>% left_join(.y))) %>%
  select(-cor_data) 

plan(multisession(workers = 12L))
mr_nested <- mr_nested %>%
  mutate(ma = 
    future_pmap(
      # pmap(
      list(data, model, emo_item, ergoal, parameter, type, o_item, exclusion, er_exc)
      , .f = possibly(mr_fun, NA_real_)
      # , mr_fun
      , .progress = T
      , .options = furrr_options(packages = c("brms", "broom.mixed", "metafor"))
    ))
closeAllConnections()
```

## Load in Results

```{r}
loadRData <- function(file, obj, rq){
  file <- sprintf("../05-results/%s/mr-fx/%s", rq, file)
  load(file)
  get(ls()[grepl(obj, ls())])
}

mr_nested <- tibble(model = c("RQ1", "RQ2-int", "RQ2-main")) %>%
  mutate(file = map(model, ~list.files(sprintf("../05-results/%s/mr-fx", .)))) %>%
  unnest(file) %>%
  filter(grepl("parameter|person", file)) %>%
  mutate(fx = pmap(list(file, "fx", model), loadRData)) %>%
  separate(file, c("emo_item", "ergoal", "parameter", "o_item", "type", "exclusion", "er_exc"), sep = "-") %>%
    mutate(
      er_exc = str_remove_all(er_exc, ".RData")
      , ergoal = ifelse(ergoal == "NA", NA_real_, ergoal)
      ) %>%
  # inner_join(mr_nested %>% select(model:er_exc)) %>%
  distinct() %>%
  filter(!is.na(exclusion)) %>%
  left_join(
    summ_df %>%
      group_by(model, emo_item, ergoal, parameter, exclusion, er_exc) %>% 
      filter(!parameter %in% c("mu_E", "sigma_E", "sigma_p")) %>%
      nest() %>%
      ungroup() %>%
      arrange(model, emo_item, ergoal, parameter) 
  ) %>%
  mutate(fx = map(fx, ~(.) %>% select(-matches("type"))))
```

## Reshape Results

```{r}
mr_fx <- mr_nested %>% 
  select(-data) %>%
  unnest(fx) %>%
  filter(
    (type == "linear" & term == "o_value") |
    (type == "quadratic" & term == "o_value_2")
    ) %>%
  mutate(
    sig = ifelse(sign(conf.low) == sign(conf.high), "sig", "ns")
    , focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")
  ) %>%
  mutate_at(vars(estimate, conf.low, conf.high), ~sprintf("%.2f", .)) %>%
  mutate(
    CI = sprintf("[%s, %s]", conf.low, conf.high)
    , est = sprintf("%s<br>[%s, %s]", estimate, conf.low, conf.high)
  ) %>%
  mutate_at(vars(est, estimate, CI), ~ifelse(sig == "sig", sprintf("<strong>%s</strong>", .), .)) %>%
  select(model, exclusion, er_exc, o_item, emo_item, ergoal, parameter, type, term, focal, est, estimate, CI)
```

```{r}
rq3_rq1_tab_fun <- function(d, exclusion, focal, er_exc){
  # h1 <- rep(1, 9); names(h1) <- c(" ", rep(c("Linear", "Quadratic"), times = 4))
  h2 <- c(2, rep(1, 4)); names(h2) <- c(" ", rep(c("Attractor Strength", "Attractor Location"), times = 2))
  h3 <- c(2, rep(2, 2)); names(h3) <- c(" ", "Negative Emotion", "Positive Emotion")
  # h2 <- c(2, rep(2, 4)); names(h2) <- c(" ", rep(c("Attractor Strength", "Attractor Location"), times = 2))
  # h3 <- c(2, rep(4, 2)); names(h3) <- c(" ", "Negative Emotion", "Positive Emotion")
  
  tab <- d %>%
    select(
      o_item, type, matches("Negative_beta")
      , matches("Negative_loc"), matches("Positive_beta")
      , matches("Positive_loc")
      ) %>%
    kable(
      .
      , "html"
      , escape = F
      # , col.names = c("Well-Being/ER Indicator", "Term", rep(c("Est.", "[CI]"), times = 4))
      , col.names = c("Well-Being/ER Indicator", "Term", rep("Est. [CI]", times = 4))
      , align = c("l", "l", rep("c", 4))
      # , align = c("l", "l", rep("c", 8))
      , cap = "<strong>Table X</strong><br><em>Linear and Quadratic Associations Between Emotion Regulation and Well-Being Indicators and Emotion Attractor Location and Strength</em>"
    ) %>%
    kable_classic(full_width = F, html_font = "Times") %>%
    # add_header_above(h1) %>%
    add_header_above(h2) %>%
    add_header_above(h3)
  save_kable(tab, file = sprintf("../05-results/tables/tab-x-rq3b-emo-wb-correlates-%s-%s-%s.html", focal, exclusion, er_exc))  
  return(tab)
}

rq3b_1_tab <- mr_fx %>%
  filter(model == "RQ1") %>%
  select(-ergoal) %>%
  # select(-term, -model, -est) %>%
  select(-term, -model, -estimate, -CI) %>%
  pivot_wider(
    names_from = c("emo_item", "parameter")
    # , values_from = c("estimate", "CI")
    , values_from = c("est")
    , names_glue = "{emo_item}_{parameter}_{.value}"
  ) %>%
  mutate(o_item = factor(o_item, cor_items$old_name, cor_items$long_name)) %>%
  arrange(exclusion, er_exc, focal, o_item, type) %>%
  group_by(exclusion, er_exc, focal) %>%
  nest() %>%
  ungroup() %>%
  mutate(tab = pmap(list(data, exclusion, focal, er_exc), rq3_rq1_tab_fun))

rq3b_1_tab %>% filter(focal == "focal" & exclusion == "parameter" & er_exc == "all") %>% pull(tab)
```

```{r}
rq3_rq2_tab_fun <- function(d, exclusion, focal, er_exc){
  h2 <- c(2, rep(2, 4)); names(h2) <- c(" ", rep(c("Attractor Strength", "Attractor Location"), times = 2))
  h3 <- c(2, rep(4, 2)); names(h3) <- c(" ", "Negative Emotion", "Positive Emotion")
  rs <- d %>% group_by(o_item) %>% tally() %>% 
        mutate(end = cumsum(n), start = lag(end) + 1, start = ifelse(is.na(start), 1, start))
  
  tab <- d %>%
    select(-o_item) %>%
    kable(
      .
      , "html"
      , escape = F
      , col.names = c("Direction", "Term", rep(c("Est.", "[CI]"), times = 4))
      , align = c("l", "l", rep("c", 8))
      , cap = "<strong>Table X</strong><br><em>Linear and Quadratic Associations Between Emotion Regulation and Well-Being Indicators and Emotion Attractor Location and Strength</em>"
    ) %>%
    kable_classic(full_width = F, html_font = "Times") %>%
    # add_header_above(h1) %>%
    add_header_above(h2) %>%
    add_header_above(h3)
  
  for (i in 1:nrow(rs)){
      tab <- tab %>% 
        kableExtra::group_rows(rs$o_item[i], rs$start[i], rs$end[i]) 
  }
  save_kable(tab, file = sprintf("../05-results/tables/tab-x-rq3b-int-wb-correlates-%s-%s-%s.html", focal, exclusion, er_exc))
  return(tab)
}

rq3b_2_tab <- mr_fx %>%
  filter(
    model == "RQ2-int" & 
    (grepl("location_", parameter) | grepl("strength_", parameter))
    ) %>%
  separate(parameter, c("parameter", "direction"), sep = "_") %>% 
  select(-term, -model, -est, -ergoal) %>%
  pivot_wider(
    names_from = c("emo_item", "parameter")
    , values_from = c("estimate", "CI")
    , names_glue = "{emo_item}_{parameter}_{.value}"
  ) %>%
  select(o_item, focal, exclusion, er_exc, direction, type, matches("Negative_str"), matches("Negative_loc"), matches("Positive_str"), matches("Positive_loc")) %>%
  mutate(
    o_item = factor(o_item, levels = cor_items$old_name, labels = cor_items$long_name)
    , direction = factor(str_to_title(direction), c("Decrease", "Maintain", "Increase"))
    , type = str_to_title(type)
    ) %>%
  arrange(o_item, direction, type) %>%
  group_by(focal, exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  mutate(tab = pmap(list(data, exclusion, focal, er_exc), rq3_rq2_tab_fun))
```

# Averages and Individual differences in Attractor Location & Strength

## Just Emotions

```{r}
summ_rq1 <- summ_df %>%
  filter(
    model == "RQ1" & 
      emo_item %in% c("Negative", "Positive") & 
      parameter %in% c("beta_E", "location") 
    ) %>%
  group_by(emo_item, parameter, exclusion, er_exc) %>%
  rename(est = mean) %>%
  summarize(
    mean   = mean(est,na.rm=T),
    sd     = sd(est,na.rm=T),
    min    = min(est,na.rm=T),
    max    = max(est,na.rm=T),
    median = median(est,na.rm=T),
    iqr    = IQR(est,na.rm=T)
  ) %>%
  ungroup() %>%
  mutate(
    parameter = factor(
      parameter, c("beta_E", "location"), 
      c("Attractor Strength", "Attractor Location")
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    )
summ_rq1
```

### Plot

```{r}
plot_rq1_dist <- function(d, exc, er_exc){
  p <- d %>%
     ggplot(aes(x = mean)) + 
      geom_histogram(aes(fill = emo_item), alpha = .5, color = "black") + 
      geom_text(
        data = summ_rq1 %>% 
          filter(exclusion == exc) %>%
          mutate(pos = ifelse(parameter == "Attractor Location", 0, -5.5))
        , aes(
          label = sprintf("M = %.2f\n(SD = %.2f)", median, sd)
          , x = pos, y = 60
          )
          , hjust = 0, vjust = 1
        ) + 
      scale_fill_manual(values = c("red3", "goldenrod2")) + 
      labs(
        x = "Parameter Estimate"
        , y = "Density"
        , title = "Individual Differences in Parameter Estimates"
      ) + 
      facet_grid(emo_item ~ parameter, scales = "free_x") + 
      my_theme() + 
     theme(legend.position = "none")
  ggsave(p, file = sprintf("../05-results/plots/RQ1/dist/rq1-comb-dist-%s-%s.png", exc), width = 10, height = 10)
  ggsave(p, file = sprintf("../05-results/plots/RQ1/dist/rq1-comb-dist-%s-%s.pdf", exc), width = 10, height = 10)
  return(p)
}

rq1_dist <- summ_df %>%
  filter(
    model == "RQ1" & 
    emo_item %in% c("Negative", "Positive") & 
    parameter %in% c("beta_E", "location")  
    ) %>%
  mutate(
    parameter = factor(
      parameter, c("beta_E", "location"), 
      c("Attractor Strength", "Attractor Location")
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    ) %>%
  group_by(exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  mutate(p = pmap(list(data, exclusion, er_exc), plot_rq1_dist))

rq1_dist$p[[2]]
```

## Main effect ER Goals

```{r}
summ_rq2 <- summ_df %>%
  filter(model == "RQ2-main" & emo_item %in% c("Negative", "Positive") & parameter %in% c("beta_E", "location", "beta_ER")) %>%
  group_by(emo_item, ergoal, parameter, exclusion, er_exc) %>%
  rename(est = mean) %>%
  summarize(
    mean   = mean(est,na.rm=T),
    sd     = sd(est,na.rm=T),
    min    = min(est,na.rm=T),
    max    = max(est,na.rm=T),
    median = median(est,na.rm=T),
    iqr    = IQR(est,na.rm=T)
  ) %>%
  ungroup() %>%
  mutate(
    parameter = factor(
      parameter, c("beta_E", "beta_ER", "location"), 
      c("Attractor Strength", "Emotion Regulation Efficacy", "Attractor Location")
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    )
summ_rq2
```

### Plots

```{r}
rq2_main_dist <- function(d, er_goal, exc, er_exc){
  eg <- mapvalues(er_goal, c("goalNeg", "goalPos"), c("Goal to Change Negative Emotions", "Goal to Change Positive Emotions"))
  p <- d %>%
    ggplot(aes(x = mean)) + 
      geom_histogram(aes(fill = emo_item), alpha = .5, color = "black") + 
      geom_text(
        data = summ_rq2 %>% 
          filter(ergoal == er_goal & exclusion == exc) %>%
          mutate(pos = 
                   ifelse(parameter == "Attractor Location", 0, 
                   ifelse(parameter == "Attractor Strength", -10, -1.5))
                 )
        , aes(
          label = sprintf("M = %.2f\n(SD = %.2f)", median, sd)
          , x = pos, y = 80
          )
          , hjust = 0, vjust = 1
        ) +
      scale_fill_manual(values = c("red3", "goldenrod2")) + 
      labs(
        x = "Individual Differences in Parameter Estimates"
        , y = "Density"
        , title = eg
        , subtitle = "Main Effect of Emotion Regulation Goal Models"
      ) + 
      facet_grid(emo_item ~ parameter, scales = "free_x") + 
      my_theme() + 
     theme(legend.position = "none")
  ggsave(p, file = sprintf("../05-results/plots/RQ2-main/dist/rq2-comb-dist-%s-%s-%s.png", ergoal, exc, er_exc), width = 10, height = 4)
  ggsave(p, file = sprintf("../05-results/plots/RQ2-main/dist/rq2-comb-dist-%s-%s-%s.pdf", ergoal, exc, er_exc), width = 10, height = 4)
  return(p)
}

rq2_main_dist_plots <- summ_df %>%
  filter(model == "RQ2-main" & emo_item %in% c("Negative", "Positive") & parameter %in% c("beta_E", "location", "beta_ER")) %>%
  mutate(
    parameter = factor(
      parameter, c("beta_E", "beta_ER", "location"), 
      c("Attractor Strength", "Emotion Regulation Efficacy", "Attractor Location")
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    ) %>%
  group_by(ergoal, exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  mutate(p = pmap(list(data, ergoal, exclusion, er_exc), rq2_main_dist))
 rq2_main_dist_plots$p[[3]] / rq2_main_dist_plots$p[[4]] + labs(subtitle = NULL)
```

## Interaction effect ER Goals on Change in Emotion

```{r}
summ_rq2_int <- summ_df %>%
  filter(model == "RQ2-int" & emo_item %in% c("Negative", "Positive") & parameter %in% c("beta_E", "mu_E", "beta_ER", "beta_ExER",   "location_decrease" , "location_maintain"  , "location_increase",
       "strength_decrease",
       "strength_maintain",
       "strength_increase")) %>%
  filter(abs(mean) < 10) %>%
  group_by(emo_item, ergoal, parameter, exclusion, er_exc) %>%
  rename(est = mean) %>%
  summarize(
    mean   = mean(est,na.rm=T),
    sd     = sd(est,na.rm=T),
    min    = min(est,na.rm=T),
    max    = max(est,na.rm=T),
    median = median(est,na.rm=T),
    iqr    = IQR(est,na.rm=T)#,
    # n      = n()
  ) %>%
  ungroup() %>%
  mutate(
    parameter = factor(
      parameter, c("beta_E", "beta_ER", "beta_ExER", "mu_E","location_decrease" , "location_maintain"  , "location_increase",
       "strength_decrease",
       "strength_maintain",
       "strength_increase"), 
      c("Attractor Strength", "Emotion Regulation Efficacy", "Emotion x Goals Interaction", "Emotion Intercept", "Location Goal Decrease" , "Location Goal Maintain"  , "Location Goal Maintain",
       "Attractor Strength Goal Decrease",
       "Attractor Strength Goal Maintain",
       "Attractor Strength Goal Increase" )
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    )
summ_rq2_int
```

### Plots

```{r}
rq2_int_dist <- function(d, er_goal, exc, er_exc){
  eg <- mapvalues(er_goal, c("goalNeg", "goalPos"), c("Goal to Change Negative Emotions", "Goal to Change Positive Emotions"))
  p <- d %>%
    ggplot(aes(x = mean)) + 
    geom_histogram(aes(fill = emo_item), alpha = .5, color = "black") + 
    geom_text(
      data = summ_rq2_int %>%
        filter(ergoal == er_goal & exclusion == exc) %>%
        mutate(pos = 
          ifelse(parameter == "Emotion Intercept", 0, 
          ifelse(parameter == "Attractor Strength", -5,
          ifelse(parameter == "Emotion x Goals Interaction", -1.5, -3)))
          )
      , aes(
        label = sprintf("M = %.2f\n(SD = %.2f)", median, sd)
        , x = pos, y = 100
        )
        , hjust = 0, vjust = .6
      ) +
    scale_fill_manual(values = c("red3", "goldenrod2")) + 
    labs(
      x = "Individual Differences in Parameter Estimates"
      , y = "Count"
      , title = eg
      , subtitle = "Emotion x Emotion Regulation Goal Interaction Models" 
    ) + 
    facet_wrap(~parameter, scales = "free_x", nrow = 2) + 
    my_theme() + 
   theme(legend.position = "none")
  ggsave(p, file = sprintf("../05-results/plots/RQ2-int/dist/rq2-comb-dist-%s-%s-%s.png", ergoal, exc, er_exc), width = 14, height = 8)
  ggsave(p, file = sprintf("../05-results/plots/RQ2-int/dist/rq2-comb-dist-%s-%s-%s.pdf", ergoal, exc, er_exc), width = 14, height = 8)
  return(p)
}

rq2_int_dist_plots <- summ_df %>%
  filter(model == "RQ2-int" & emo_item %in% c("Negative", "Positive") & !grepl("sigma", parameter)) %>%
  mutate(
    parameter = factor(
      parameter, c("beta_E", "beta_ER", "beta_ExER", "mu_E","location_decrease" , "location_maintain"  , "location_increase",
       "strength_decrease",
       "strength_maintain",
       "strength_increase"), 
      c("Attractor Strength", "Emotion Regulation Efficacy", "Emotion x Goals Interaction", "Emotion Intercept", "Location Goal Decrease" , "Location Goal Maintain"  , "Location Goal Maintain",
       "Attractor Strength Goal Decrease",
       "Attractor Strength Goal Maintain",
       "Attractor Strength Goal Increase" )
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    ) %>%
  group_by(ergoal, exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  mutate(p = pmap(list(data, ergoal, exclusion, er_exc), rq2_int_dist))
rq2_int_dist_plots$p[[3]] / rq2_int_dist_plots$p[[4]] + labs(subtitle = NULL)

```

## Caterpillar Plots

```{r}
cat_plot_fun <- function(d){
  ord <- d %>% 
    arrange(desc(mean)) %>%
    pull(SID)
  x %>%
    ggplot(aes(x = value, y = id)) + 
      geom_line(aes(group = id)) + 
      geom_point(aes(color = type)) +
      scale_color_manual(values = c("blue", "red")) + 
      labs(x = par, y = NULL, title = par) + 
      my_theme()
  p1 <- d %>%
    mutate(
      SID = factor(SID, levels = ord)
      , sig = ifelse(sign(lower) == sign(upper), "sig", "ns")
      ) %>%
    ggplot(aes(x = mean, y = SID)) + 
      # geom_density(aes(y = after_stat(density))) + 
      geom_errorbar(
        aes(xmin = lower, xmax = upper)
        , width = .1
      ) + 
      geom_point() +
      labs(
        x = "Parameter Estimate"
        , y = "SID"
        , title = "Individual Differences in Parameter Estimates"
      ) + 
      # facet_grid(emo_item ~ parameter, scales = "free_x") + 
      my_theme() + 
      theme(axis.text.y = element_blank())
  
  d %>%
    ggplot(aes(x = mean)) + 
      geom_density() + 
      theme_void()
}

summ_df %>%
  filter(
    model == "RQ1" & 
    emo_item %in% c("Negative", "Positive") & 
    parameter %in% c("beta_E", "location")
    ) %>%
  group_by(model, emo_item, ergoal, parameter) %>%
  nest() %>%
  ungroup() 
  # mutate(
  #   parameter = factor(
  #     parameter, c("beta_E", "location"), 
  #     c("Attractor Strength", "Attractor Location")
  #     )
  #   , emo_item = factor(
  #     emo_item, c("Negative", "Positive")
  #     , c("Negative Emotion", "Positive Emotion")
  #     )
  #   ) %>%
  
   
```

# RQ3 - Correlates

```{r}
cor_tab_fun <- function(r){
  r$r <- apply(r$r, c(1,2), function(x) sprintf("%.2f", x))
  r$r[which(r$p < .05, arr.ind = T)] <- 
    sprintf("<strong>%s</strong>" ,r$r[which(r$p < .05, arr.ind = T)])
  r$r[upper.tri(r$r, diag = F)] <- ""
  diag(r$r) <- "--"
  r <- r$r %>% data.frame %>%
    rownames_to_column("item") %>%
    as_tibble() %>%
    mutate_all(as.character) %>%
    select(item, everything()) %>%
    mutate(item = sprintf("%i. %s", seq(1, ncol(r$r)), item)) %>%
    setNames(c("item", seq(1, ncol(r$r)))) 
  return(r)
}
```

## RQ3a - Correlations among parameters

```{r}
nested_r <- summ_df %>% 
  filter(model %in% c("RQ1", "RQ2-int") & grepl("strength|location", parameter)) %>%
  # filter((parameter == "location" & mean > 0) | parameter == "beta_E") %>%
  select(model, SID, emo_item, ergoal, parameter, mean, exclusion, er_exc) %>%
  # mutate(parameter = mapvalues(parameter, c("beta_E"), c("strength"))) %>%
  mutate(focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")) %>%
  group_by(model, focal, exclusion, er_exc) %>%
  nest() %>%
  ungroup()
  # mutate(data = map(data, ~(. %>% 
  #     pivot_wider(
  #       names_from = c("emo_item", "parameter")
  #       , values_from = "mean"
  #     )



cor_tab_html_fun <- function(tab, focal, rq, exc, er_exc){
  len <- nrow(tab)
  tab <- tab %>%
    separate(item, c("num", "emo_item", "param"), sep = "[ ]|_") %>%
    kable(
      .
      , "html"
      , escape = F
      , col.names = c(" ", "Variable", "Parameter", paste0(1:len, "."))
      , align = c("l", "l", "l", rep("c", len))
      , caption = sprintf("<strong>Table SX</strong><br><em>RQ3a: Correlations Among %s Parameters</em>", rq)
    ) %>%
    kable_classic(html_font = "Times")
  save_kable(tab, file = sprintf("../05-results/tables/%s/correlations/%s-%s-%s.html", rq, focal, exc, er_exc))
  return(tab)
}

lm_fun <- function(d) tidy(effectsize::standardize(lm(strength ~ location + location2, data = d)), conf.int = T)

quad_fun <- function(d){
  d %>% 
    pivot_wider(
      names_from = "parameter"
      , values_from = "mean"
    ) %>%
    mutate(location2 = location^2) %>%
    group_by(emo_item) %>%
    nest() %>%
    ungroup() %>%
    mutate(m = map(data, lm_fun)) %>%
    select(-data) %>% 
    unnest(m) %>%
    filter(term == "location2")
}

nested_rq1_r <- summ_df %>% 
  filter(model == "RQ1" & parameter %in% c("beta_E", "location")) %>%
  filter((parameter == "location" & mean > 0) | parameter == "beta_E") %>%
  select(SID, exclusion, er_exc, emo_item, parameter, mean) %>%
  mutate(
    focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")
    , parameter = mapvalues(parameter, c("beta_E"), c("strength"))
    ) %>%
  arrange(focal, emo_item, parameter) %>%
  group_by(focal, exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  mutate(
    r2 = map(data, quad_fun)
    , data = map(data, ~(.) %>% 
      pivot_wider(
        names_from = c("emo_item", "parameter")
        , values_from = "mean"
      ))
    , r = map(data, ~(.) %>% select(-SID) %>% corr.test())
    , rtab = map(r, cor_tab_fun)
    , rtab_html = pmap(list(rtab, focal, "RQ1", exclusion, er_exc), cor_tab_html_fun)
  )

nested_rq1_r %>% filter(focal == "focal") %>% pull(rtab_html)

rq1_r2 <- nested_rq1_r %>% 
  select(focal, r2, exclusion, er_exc) %>%
  unnest(r2) %>%
  select(-term) %>%
  mutate_at(vars(estimate, conf.low, conf.high), ~sprintf("%.2f", .)) %>%
  mutate(
    sig = ifelse(p.value < .05, "sig", "ns")
    , CI = sprintf("[%s, %s]", conf.low, conf.high)
    ) %>%
  mutate_at(vars(estimate, CI), ~ifelse(sig == "sig", sprintf("<strong>%s</strong>", .), .)) %>%
  select(focal, emo_item, estimate, CI) %>%
  mutate(focal = factor(focal, c("focal", "sup"), c("Focal", "Supplementary"))) %>%
  kable(
    .
    , "html"
    , escape = F
    , col.names = c("Emotion Item Set", "Emotion", "Std. Est.", "CI")
    , align = c("l", "l", "c", "c")
    , cap = "<strong>Table SX</strong><br>Associations between Attractor Location^2 and Attractor Strength from RQ1"
  ) %>%
  kable_classic(full_width = F, html_font = "Times") %>%
  collapse_rows(1, "top")
rq1_r2
save_kable(rq1_r2, file = "../05-results/tables/tab-sx-loc2-cors.html")
```

```{r}
cor_tab_html_fun <- function(tab, focal, rq, exclusion, er_exc){
  tab <- tab %>%
    separate(
      item
      , c("num", "emo_item", "ergoal", "parameter", "ergoal_lev")
      , sep = "[ ]|_")
  len <- nrow(tab)
  tab <- tab %>%
    select(-emo_item, -ergoal) %>%
    kable(
      .
      , "html"
      , escape = F
      , col.names = c(" ", "Variable", " ", paste0(1:len, "."))
      , align = c("l", rep("c", len + 2))
      , caption = sprintf("<strong>Table SX</strong><br><em>RQ3a: Correlations Among %s Parameters</em>", rq)
    ) %>%
    kable_classic(html_font = "Times") %>%
    kableExtra::group_rows("Negative Emotions: Change Negative Emotions", 1, len/2) %>%
    kableExtra::group_rows("Positive Emotions: Change Positive Emotions", len/2+1, len) %>%
    add_header_above(c(" " = 3, "Negative Emotions: Change Negative Emotions" = len/2, "Positive Emotions: Change Positive Emotions" = len/2)) %>%
    add_header_above(c(" " = 3, "Zero-Order Correlations" = len))
  save_kable(tab, file = sprintf("../05-results/tables/%s/correlations/%s-%s-%s.html", rq, focal, exclusion, er_exc))
  return(tab)
}
nested_rq2int_r <- summ_df %>% 
  filter(model == "RQ2-int" & grepl("location_|strength_", parameter)) %>%
  select(SID, exclusion, er_exc, emo_item, ergoal, parameter, mean) %>%
  filter((ergoal == "goalNeg" & emo_item %in% c("Negative")) | 
         (ergoal == "goalPos" & emo_item %in% c("Positive"))) %>%
  mutate(
    focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")
    ) %>%
  arrange(focal, emo_item, ergoal, parameter) %>%
  group_by(focal, exclusion, er_exc) %>%
  nest() %>%
  mutate(
    data = map(data, ~(.) %>% 
      pivot_wider(
        names_from = c("emo_item", "ergoal", "parameter")
        , values_from = "mean"
      ))
    , r = map(data, ~(.) %>% select(-SID) %>% corr.test())
    , rtab = map(r, cor_tab_fun)
    , rtab_html = pmap(list(rtab, focal, "RQ2-int", exclusion, er_exc), cor_tab_html_fun)
  )

# nested_rq2int_r <- summ_df %>% 
#   filter(model == "RQ2-int" & grepl("location_|strength_", parameter)) %>%
#   select(SID, exclusion, emo_item, ergoal, parameter, mean) %>%
#   filter((ergoal == "goalNeg" & emo_item %in% c("Negative", neg_emo_items$long_name)) | 
#          (ergoal == "goalPos" & emo_item %in% c("Positive", pos_emo_items$long_name))) %>%
#   mutate(
#     focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")
#     ) %>%
#   arrange(focal, emo_item, ergoal, parameter) %>%
#   group_by(focal, exclusion) %>%
#   nest() %>%
#   mutate(
#     data = map(data, ~(.) %>% 
#       pivot_wider(
#         names_from = c("emo_item", "ergoal", "parameter")
#         , values_from = "mean"
#       ))
#     , r = map(data, ~(.) %>% select(-SID) %>% corr.test())
#     , rtab = map(r, cor_tab_fun)
#     , rtab_html = pmap(list(rtab, focal, "RQ2-int"), cor_tab_html_fun)
#   )

nested_rq2int_r$rtab_html
```

# Getting at Goal x Emotion Dynamics

```{r}
View(summ_df %>% 
  filter(parameter == "beta_ExER") %>%
  mutate(
    sign = sign(mean)
    , sig = ifelse(sign(lower) == sign(upper), "sig", "ns")
    ) %>%
  group_by(sign, sig, emo_item, ergoal, exclusion) %>% 
  tally() %>% 
  ungroup() %>%
  pivot_wider(
    names_from = c("sign", "sig")
    , values_from = "n"
    ) %>%
  group_by(emo_item, ergoal, exclusion) %>%
  mutate(
    ns = sum(`-1_ns`, `1_ns`)
    , total = sum(across(`-1_ns`:`1_sig`))
    ) %>%
  mutate_at(vars(`-1_ns`:ns), ~sprintf("%i (%.2f%%)", ., ./total*100)) %>%
  arrange(exclusion, ergoal, emo_item))
```

## 4. Classifying direction of interaction term & main effect of goal

```{r}
rq2_sum_negintmain<-summ_df %>%
   filter(.,model=="RQ2-int",emo_item=="Negative",ergoal=="goalNeg",parameter %in% c("beta_ExER", "beta_ER"),abs(mean) < 10)%>%
  pivot_wider(names_from = parameter, values_from =c(mean,sd,lower,mid,upper))%>%
  mutate(int_pos_m_pos = ifelse(mean_beta_ExER > 0 & lower_beta_ExER > 0 &
                                  upper_beta_ExER > 0 & mean_beta_ER > 0 &
                                  lower_beta_ER > 0 & upper_beta_ER > 0, 1, 0),
          int_pos_m_neg = ifelse(mean_beta_ExER > 0 & lower_beta_ExER > 0 & upper_beta_ExER > 0
                                 & mean_beta_ER < 0 & lower_beta_ER < 0 & upper_beta_ER < 0 ,1,
                                 0),
         int_neg_m_neg = ifelse(mean_beta_ExER < 0 & lower_beta_ExER < 0 & upper_beta_ExER < 0
                                & mean_beta_ER < 0 & lower_beta_ER < 0 & upper_beta_ER < 0 , 1,
                                0),
         int_neg_m_pos = ifelse(mean_beta_ExER < 0 & lower_beta_ExER < 0 & upper_beta_ExER < 0
                                & mean_beta_ER > 0 & lower_beta_ER > 0 & upper_beta_ER > 0 , 1,
                                0))%>%
           mutate(int_ns_m_pos = ifelse(int_pos_m_pos==0 & int_pos_m_neg==0 & 
                                          int_neg_m_pos==0 & int_neg_m_neg==0 & 
                                          
                                          mean_beta_ER > 0 & lower_beta_ER > 0 & upper_beta_ER
                                        > 0 , 1, 0),
                  int_ns_m_neg = ifelse(int_pos_m_pos==0 & int_pos_m_neg==0 & 
                                          int_neg_m_pos==0 & int_neg_m_neg==0 & 
                                          mean_beta_ER < 0 & lower_beta_ER < 0 & upper_beta_ER
                                        < 0, 1, 0),
                  int_pos_m_ns = ifelse(int_pos_m_pos==0 & int_pos_m_neg==0 & 
                                          int_neg_m_pos==0 & int_neg_m_neg==0 & 
                                          mean_beta_ExER > 0 & lower_beta_ExER > 0 &
                                          upper_beta_ExER > 0, 1, 0),
                  int_neg_m_ns = ifelse(int_pos_m_pos==0 & int_pos_m_neg==0 & 
                                          int_neg_m_pos==0 & int_neg_m_neg==0 & 
                                          mean_beta_ExER < 0 & lower_beta_ExER < 0 &
                                          upper_beta_ExER < 0, 1, 0))%>%
  
  mutate(int_ns_m_ns = ifelse(int_pos_m_pos==0 & int_pos_m_neg==0 &
                                int_neg_m_pos==0 & int_neg_m_neg==0 &
                                int_ns_m_pos==0 & int_ns_m_neg == 0 &
                                int_pos_m_ns ==0 & int_neg_m_ns==0, 1, 0))%>%
  
  select(SID,starts_with("int"))%>%
   rename_with(~paste("Negative", ., sep = "_"), starts_with("int"))

describe(rq2_sum_negintmain)

View(filter(rq2_sum_negintmain,SID=="15604"))
```

## 5. Time that people were successful

```{r}
loadRData <- function(file, obj, rq){
  file <- sprintf("../05-results/%s/success/%s", rq, file)
  load(file)
  get(ls()[grepl(obj, ls())])
}

nested_perc_data <- tibble(model = c("RQ2-int","RQ2-main")) %>%
  mutate(file = map(model, ~list.files(sprintf("../05-results/%s/success", .)))) %>%
  unnest(file) %>%
  mutate(perc_data = pmap(list(file, "perc_data", model), loadRData)) %>% 
  separate(file, c("study", "SID", "emo_item", "ergoal"), sep = "-") %>%
  mutate_at(vars(emo_item, ergoal), ~str_remove_all(., ".RData"))

pos_perc_data<-
  nested_perc_data %>% 
  unnest(perc_data) %>% 
  filter(emo_item=="Positive")%>%
  mutate(er_goal = recode(er_goal, "-2.5" = "de", "0" = "mt", "2.5" = "in"),zone=recode(zone, "Successful"="suc","Unsuccessful"="unsuc",Counterproductive="count"))%>%
  filter(model == "RQ2-int") %>%
  pivot_wider(
    names_from = c("emo_item","zone","er_goal")
    , values_from = "n"
  ) 

neg_perc_data<-
  nested_perc_data %>% 
  unnest(perc_data) %>% 
  filter(emo_item=="Negative")%>%
  mutate(er_goal = recode(er_goal, "-2.5" = "de", "0" = "mt", "2.5" = "in"),zone=recode(zone, "Successful"="suc","Unsuccessful"="unsuc",Counterproductive="count"))%>%
  filter(model == "RQ2-int") %>%
  pivot_wider(
    names_from = c("emo_item","zone","er_goal")
    , values_from = "n"
  ) 

# adding 0 for goals  people are predicted to never be successful/unsuccessful/counterproductive

pos_perc_data[is.na(pos_perc_data)] <- 0
neg_perc_data[is.na(neg_perc_data)] <- 0

erstrat$SID<-as.character(erstrat$SID)

corrdat5<-left_join(select(pos_perc_data,!c("ergoal","model")),select(neg_perc_data,!c("ergoal","model")),by="SID")%>%
  left_join(.,wb_wide)%>%
    left_join(.,select(erstrat,!"StartDate"))

View(corrdat5)

# Well-being outcomes
corr.test(select(corrdat5,c("Positive_unsuc_de", "Positive_suc_mt", "Positive_unsuc_mt", 
"Positive_suc_in", "Positive_unsuc_in", "Positive_suc_de", "Positive_count_in", 
"Positive_count_de", "Negative_suc_de", "Negative_unsuc_de", 
"Negative_suc_mt", "Negative_unsuc_mt", "Negative_suc_in", "Negative_unsuc_in", 
"Negative_count_de", "Negative_count_in", "DEP", "SWL")))

# depression
# positive decrease successful r = -.18 (higher depression, less time decreasing positive emotions successfully)
# negative increase successful r = -.21 (higher depression less time increasing negative emotions successfully)
# negative decrease counterproductive r = .29 (higher depression more time that negative emotions actually increase when they want them to decrease)
#  negative increase counterproductive r =.28 (higher depression more time that negative emotions actually decrease when they want them to increase)

# SWL
# negative decrease counterproductive r = -.20 (higher SWl, less time counterproductive decreasing of negative affect)



corr.test(select(corrdat5,c("Positive_unsuc_de", "Positive_suc_mt", "Positive_unsuc_mt", 
"Positive_suc_in", "Positive_unsuc_in", "Positive_suc_de", "Positive_count_in", 
"Positive_count_de", "Negative_suc_de", "Negative_unsuc_de", 
"Negative_suc_mt", "Negative_unsuc_mt", "Negative_suc_in", "Negative_unsuc_in", 
"Negative_count_de", "Negative_count_in","er_strat_rep")))

# Positive_unsuc_mt - r =-.18, lower repertoire 
# Positive_suc_de - r =-.46 lower repertoire
# Positive_count_in - r = -.23 lower repertoire
# Negative_suc_de - r =-.28 lower repertoire
# Negative_suc_in - r = -0.52 lower repertoire


corr.test(select(corrdat5,c("Positive_unsuc_de", "Positive_suc_mt", "Positive_unsuc_mt", 
"Positive_suc_in", "Positive_unsuc_in", "Positive_suc_de", "Positive_count_in", 
"Positive_count_de", "Negative_suc_de", "Negative_unsuc_de", 
"Negative_suc_mt", "Negative_unsuc_mt", "Negative_suc_in", "Negative_unsuc_in", 
"Negative_count_de", "Negative_count_in", "avg_accept", "avg_coRea", 
"avg_dist", "avg_mask", "avg_min", "avg_posRea", "avg_rem", "avg_rum", 
"avg_savor", "avg_selApp", "avg_selAv", "avg_socShr", "avg_sup", 
"avg_talkPos")))


## acceptance 
# pos suc mt r=  0.17 
# pos unsuc mt r =  -0.23
# Positive_count_in r =-0.25
# negative unsuc de r = -0.21

## coRea
# nothing

##distraction
#Negative_unsuc_de  r=-0.19
#Negative_unsuc_in r = -0.22

##mask
#Positive_unsuc_mt 
# Negative_unsuc_de
# Negative_unsuc_in

##pos reapp
# Positive_unsuc_mt 

## reminiscing
# nothing

# rumination
# Negative_unsuc_in

## savoring
# nothing

##selection approach or avoid 
# nothing 


## social sharing
# positive succ mt
# positive unsuc mt 
#Positive_suc_de
# Negative_suc_de

## suppression
# Positive_unsuc_mt r= -0.19
# Negative_unsuc_de r =  -0.26
# Negative_unsuc_in r =  -0.23

## talk pos - capitalization
# Negative_suc_de
# Negative_suc_mt


# do one version with combining counter and unsuccessful 

View(corrdat5)

View(select(pos_perc_data,c("Positive_suc_in", "Positive_unsuc_in", "Positive_count_in")))
       
dput(colnames(pos_perc_data)) 
```

## 6. Location & Strength when considering Goals

```{r}
rq2_corr_dat <- summ_df %>% 
  filter(model == "RQ2-int" & parameter %in% c("location_decrease" , "location_maintain"  , "location_increase",
       "strength_decrease",
       "strength_maintain",
       "strength_increase")) %>%
  filter((ergoal=="goalNeg"&emo_item=="Negative")|(ergoal=="goalPos"&emo_item=="Positive"))%>%
  select(SID, emo_item, parameter, mean)%>%
  pivot_wider(
    names_from = c("emo_item", "parameter")
    , values_from = "mean"
  )%>%
  full_join(
    wb_long %>%
      pivot_wider(names_from = wb_item, values_from =wb_value) %>% 
      mutate(SID = as.character(SID)) %>%
      select(SID, DEP, SWL)
  ) %>%
  left_join(
    erstrat %>% 
      select(SID, er_strat_rep:avg_talkPos) %>%
      mutate(SID = as.character(SID))
    )

dput(colnames(rq2_corr_dat))

summ_rq2 <- summ_df %>%
  filter(model == "RQ2-int" & emo_item %in% c("Negative", "Positive") & parameter %in% c( "strength_decrease",
       "strength_maintain",
       "strength_increase")) %>%
  group_by(emo_item, parameter) %>%
  rename(est = mean) %>%
  summarize(
    mean   = mean(est,na.rm=T),
    sd     = sd(est,na.rm=T),
    min    = min(est,na.rm=T),
    max    = max(est,na.rm=T),
    median = median(est,na.rm=T),
    iqr    = IQR(est,na.rm=T)
  ) %>%
  ungroup() %>%
  mutate(
    parameter = factor(
      parameter, c( "strength_decrease",
       "strength_maintain",
       "strength_increase"), 
      c("Decrease",
       "Maintain",
       "Increase")
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    )

#View(summ_rq2)

summ_df %>%
  filter(model == "RQ2-int" & emo_item %in% c("Negative", "Positive") & parameter %in% c(
       "strength_decrease",
       "strength_maintain",
       "strength_increase")) %>%
    filter((ergoal=="goalNeg"&emo_item=="Negative")|(ergoal=="goalPos"&emo_item=="Positive"))%>%
  mutate(
    parameter = factor(
      parameter, c(
       "strength_decrease",
       "strength_maintain",
       "strength_increase"), 
      c( "Decrease",
       "Maintain",
       "Increase")
      )
    , emo_item = factor(
      emo_item, c("Negative", "Positive")
      , c("Negative Emotion", "Positive Emotion")
      )
    ) %>%
   ggplot(aes(x = mean)) + 
    geom_histogram(aes(fill = emo_item), alpha = .5, color = "black")+ #+ 
    # geom_text(
    #   data = summ_rq2 %>% mutate(pos = ifelse(parameter == "Increase", -6.5, ifelse(parameter =="Maintain",-5.5,-9.5)))
    #   , aes(
    #     label = sprintf("M = %.2f\n(SD = %.2f)", median, sd)
    #     , x = pos, y = 80
    #     )
    #     , hjust = 0, vjust = 1
    #   ) +
    scale_fill_manual(values = c("red3", "goldenrod2")) + 
    labs(
      x = "Attractor Strength"
      , y = "Density"
      , title = "Individual Differences in Attractor Strength Estimates"
    ) + 
    facet_grid(emo_item ~ parameter) + 
    my_theme() + 
   theme(legend.position = "none")


corr.test(select(rq2_corr_dat,c(, "Negative_location_maintain", "Negative_strength_decrease", 
"Negative_strength_maintain", "Negative_strength_increase", "Positive_location_maintain", 
"Positive_location_increase", "Positive_strength_decrease", "Positive_strength_maintain", 
"Positive_strength_increase", "Negative_location_decrease", "Negative_location_increase", 
"Positive_location_decrease", "DEP", "SWL"
# , "er_strat_rep", "avg_accept", 
# "avg_coRea", "avg_dist", "avg_mask", "avg_min", "avg_posRea", 
# "avg_rem", "avg_rum", "avg_savor", "avg_selApp", "avg_selAv", 
# "avg_socShr", "avg_sup", "avg_talkPos"
)), use = "pairwise")


corr.test(select(rq2_corr_dat,c( "Negative_strength_decrease", 
"Negative_strength_maintain", "Negative_strength_increase", "Positive_strength_decrease", "Positive_strength_maintain", 
"Positive_strength_increase", "DEP", "SWL"
# , "er_strat_rep", "avg_accept", 
# "avg_coRea", "avg_dist", "avg_mask", "avg_min", "avg_posRea", 
# "avg_rem", "avg_rum", "avg_savor", "avg_selApp", "avg_selAv", 
# "avg_socShr", "avg_sup", "avg_talkPos"
)), use = "pairwise")


corr.test(select(rq2_corr_dat,c( "Negative_strength_decrease", 
"Negative_strength_maintain", "Negative_strength_increase", "Positive_strength_decrease", "Positive_strength_maintain", 
"Positive_strength_increase", #"DEP", "SWL"
, "er_strat_rep", "avg_accept",
"avg_coRea", "avg_dist", "avg_mask", "avg_min", "avg_posRea",
"avg_rem", "avg_rum", "avg_savor", "avg_selApp", "avg_selAv",
"avg_socShr", "avg_sup", "avg_talkPos"
)), use = "pairwise")
```
