==> 04-scripts/04-plots.qmd <==
---
title: "04-plots"
format: html
editor_options: 
  chunk_output_type: console
---

# Change As Outcome Plots 

## Load Data 


```{r}
loadRData <- function(file, obj, rq){
  file <- sprintf("../05-results/%s/fx/%s", rq, file)
  load(file)
  get(ls()[grepl(obj, ls())])
}

nested_sum <- tibble(model = c("RQ1", "RQ2-int", "RQ2-main")) %>%
  mutate(file = map(model, ~list.files(sprintf("../05-results/%s/fx", .)))) %>%
  unnest(file) %>%
  mutate(
    fx = pmap(list(file, "fx", model), loadRData)
    , fx = map(fx, unnest)
    ) %>%
  separate(file, c("study", "SID", "emo_item", "ergoal", "er_exc"), sep = "-") %>%
  mutate_at(vars(emo_item, ergoal, er_exc), ~str_remove_all(., ".RData")) %>%
  mutate(
    er_exc = ifelse(is.na(er_exc), ergoal, er_exc)
    , ergoal = ifelse(ergoal %in% c("all", "er"), NA_real_, ergoal)
    )

# exclude by person
# some sort of group_by and filter(any())
summ_df <- nested_sum %>% 
  unnest(fx) %>% 
  group_by(model, study, SID, emo_item, ergoal, er_exc) %>%
  mutate(flag = ifelse(
    (model == "RQ1" & grepl("beta_E", parameter) & sign(mean) == 1) |
    (grepl("mu", parameter) & (mean > 10 | mean < 0)) |
    sd == 0 | 
    sd > 10
    , "implausible", "plausible"))
  
summ_df <- summ_df %>% 
  filter(all(flag == "plausible")) %>%
  ungroup() %>%
  mutate(exclusion = "person") %>%
  full_join(
    summ_df %>%
      filter(flag == "plausible") %>%
      mutate(exclusion = "parameter")
  )
```

```{r}
nested_dat <- tibble(file = list.files("../02-data/02-clean-data/rq1/rdata/")) %>%
  mutate(data = map(file, function(y) {load(sprintf("../02-data/02-clean-data/rq1/rdata/%s", y)); return(x)})) %>%
  separate(file, c("sample", "SID", "emo_item"), sep = "-") %>%
  mutate(
    emo_item = str_remove_all(emo_item, ".RData")
    , model = "RQ1"
    ) %>%
  full_join(
    tibble(file = list.files("../02-data/02-clean-data/rq2/rdata/")) %>%
      separate(file, c("sample", "SID", "emo_item", "ergoal"), sep = "-", remove = F) %>%
      mutate(ergoal = str_remove_all(ergoal, ".RData")) %>%
      filter((ergoal == "goalNeg" & emo_item %in% c("Negative", neg_emo_items$long_name)) | 
             (ergoal == "goalPos" & emo_item %in% c("Positive", pos_emo_items$long_name))) %>%
      mutate(data = map(file, function(y) {load(sprintf("../02-data/02-clean-data/rq2/rdata/%s", y)); return(x)})) %>%
      crossing(tibble(model = c("RQ2-main", "RQ2-int")))
  ) %>%
    mutate(er_exc = "all") %>%
  full_join(
    tibble(file = list.files("../02-data/02-clean-data/rq1-robustness/rdata/")) %>%
    mutate(data = map(file, function(y) {load(sprintf("../02-data/02-clean-data/rq1-robustness/rdata/%s", y)); return(x)})) %>%
    separate(file, c("sample", "SID", "emo_item"), sep = "-") %>%
    mutate(
      emo_item = str_remove_all(emo_item, ".RData")
      , model = "RQ1"
      ) %>%
    full_join(
      tibble(file = list.files("../02-data/02-clean-data/rq2-robustness/rdata/")) %>%
        separate(file, c("sample", "SID", "emo_item", "ergoal"), sep = "-", remove = F) %>%
        mutate(ergoal = str_remove_all(ergoal, ".RData")) %>%
        filter((ergoal == "goalNeg" & emo_item %in% c("Negative", neg_emo_items$long_name)) | 
               (ergoal == "goalPos" & emo_item %in% c("Positive", pos_emo_items$long_name))) %>%
        mutate(data = map(file, function(y) {load(sprintf("../02-data/02-clean-data/rq2-robustness/rdata/%s", y)); return(x)})) %>%
        crossing(tibble(model = c("RQ2-main", "RQ2-int"))) 
    ) %>%
    mutate(er_exc = "er")
  )
```



## Individual Differences 
### RQ1 
#### Functions 

```{r}
plot_uni_traj <- function(d, emo_item, exclusion, er_exc, colr){
  ei <- if(emo_item %in% c("Negative", "Positive")) paste(emo_item, "Emotion") else emo_item
  p <- d %>%
    ggplot(aes(x = emo_value, y = dE)) + 
    geom_hline(
      aes(yintercept = 0)
      , linetype = "dashed"
      ) + 
    geom_line(
      aes(group = SID)
      , color = colr
      , linewidth = .25
      ) + 
    annotate(
      "segment", size = 1, x = 0, xend = 0, y = .25, yend = 10
      , arrow = arrow(type = "closed", length=unit(2, "mm"))
      ) +
    annotate(
      "text", x = -.25, y = 1, angle = 90, hjust = 0
      , label = paste(ei, "is increasing")
      ) +
    annotate(
      "segment", size = 1, x = 0, xend = 0, y = -.25, yend = -10
      , arrow = arrow(type = "closed", length=unit(2, "mm"))
      ) +
    annotate(
      "text", x = -.25, y = -1, angle = 90, hjust = 1
      , label = paste(ei, "is decreasing")
      ) +
    labs(
      x = sprintf("%s State", ei)
      , y = "dE/dt"
      , title = sprintf("Simple %s Emotion System", emo_item)
      # , subtitle = "Idiographic Model Estimates"
      ) +
  ylim(-10, 10) +
  my_theme()
  p
}

comb_uni_traj <- function(p, foc, exc, er_exc){
  nr <- if(foc == "focal") 1 else 4
  nc <- 2
  wdth <- 12
  ht <- if(foc == "focal") 6 else 16
  
  p2 <- wrap_plots(p$p, ncol = nc, nrow = nr, byrow = F)
  ggsave(sprintf("../05-results/plots/RQ1/%s-%s-%s.png", foc, exc, er_exc), width = wdth, height = ht)
  return(p2)
}
```

#### Run Plots 

```{r, fig.height=6, fig.width=12}
rq1_df <- summ_df %>%
  filter(model == "RQ1") %>%
  filter(parameter %in% c("mu_E", "beta_E")) %>%
  select(SID, study, emo_item, exclusion, er_exc, parameter, mean) %>%
  pivot_wider(
    names_from = "parameter"
    , values_from = mean
  )

idio_seq <- function(d){
  rng <- range(d$emo_value)
  tibble(emo_value = seq(rng[1], rng[2], .1))
}

# plot individual panels 
p_rq1 <- rq1_df %>%
  full_join(
    nested_dat %>%
      filter(model == "RQ1") %>%
      mutate(
        idio_seq = map(data, idio_seq)
        , SID = as.character(SID)
        ) %>%
      select(SID, sample, emo_item, idio_seq, er_exc) %>%
      unnest(idio_seq)
  ) %>%
  mutate(dE = mu_E + beta_E*emo_value) %>%
  group_by(emo_item, exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  full_join(emo_items %>% rename(emo_item = long_name)) %>%
  mutate(p = pmap(list(data, emo_item, exclusion, er_exc, color), plot_uni_traj)) 

# split by focal/no focal
p_rq1_comb <- p_rq1 %>%
  mutate(focal = ifelse(emo_item %in% c("Positive", "Negative"), "focal", "sup")) %>%
  select(-data) %>%
  arrange(exclusion, category, emo_item) %>%
  group_by(exclusion, focal, er_exc) %>%
  nest() %>%
  filter(!is.na(exclusion)) %>%
  ungroup() %>% 
  mutate(p_comb = pmap(list(data, focal, exclusion, er_exc), comb_uni_traj))

(p_rq1_comb %>% filter(exclusion == "parameter" & focal == "focal"))$p_comb[[1]]
```

### RQ2 Interactions Effect 
#### Functions 

```{r}
plot_int_traj <- function(d, emo_item, ergoal, exclusion, er_exc, colr){
  ei <- if(emo_item %in% c("Negative", "Positive")) paste(emo_item, "Emotion") else emo_item
  p <- d %>%
    ggplot(aes(x = emo_value, y = dE)) + 
    geom_hline(
      aes(yintercept = 0)
      , linetype = "dashed"
      ) + 
    geom_line(
      aes(group = SID)
      , color = colr
      , linewidth = .25
      ) + 
    annotate(
      "segment", size = 1, x = 0, xend = 0, y = .25, yend = 10
      , arrow = arrow(type = "closed", length=unit(2, "mm"))
      ) +
    annotate(
      "text", x = -.25, y = 1, angle = 90, hjust = 0
      , label = paste(ei, "is increasing")
      ) +
    annotate(
      "segment", size = 1, x = 0, xend = 0, y = -.25, yend = -10
      , arrow = arrow(type = "closed", length=unit(2, "mm"))
      ) +
    annotate(
      "text", x = -.25, y = -1, angle = 90, hjust = 1
      , label = paste(ei, "is decreasing")
      ) +
    labs(
      x = sprintf("%s State", ei)
      , y = "dE/dt"
      , title = sprintf("Goal-Contingent %s Emotion System", emo_item)
      # , subtitle = "Idiographic Model Estimates"
      ) +
  ylim(-10, 10) +
  facet_grid(~er_fac) +
  my_theme()
  filestr <- paste0(c(emo_item, ergoal, exclusion, er_exc), collapse = "-")
  ggsave(p, file = sprintf("../05-results/plots/RQ2-int/ind-diff-traj/%s.pdf", filestr), width = 14, height = 6)
  ggsave(p, file = sprintf("../05-results/plots/RQ2-int/ind-diff-traj/png/%s.png", filestr), width = 14, height = 6)
  p
}
```

#### Run Plots 
```{r, fig.width=14, fig.height=12}
rq2int_df <- summ_df %>%
  filter(model == "RQ2-int") %>%
  filter(parameter %in% c("mu_E", "beta_E", "beta_ER", "beta_ExER")) %>%
  select(SID, study, emo_item, ergoal, exclusion, er_exc, parameter, mean) %>%
  pivot_wider(
    names_from = "parameter"
    , values_from = mean
  ) 

idio_seq <- function(d){
  rng <- range(d$emo_value)
  crossing(
    emo_value = seq(rng[1], rng[2], .1)
    , ergoal_value = c(-2.5, 0, 2.5)
    )
}

# load("../02-data/02-clean-data/rq2/nested_dat.RData")

p_rq2int <- rq2int_df %>%
  full_join(
    nested_dat %>%
      filter(model == "RQ2-int") %>%
      mutate(
        idio_seq = map(data, idio_seq)
        , SID = as.character(SID)
        ) %>%
      select(SID, sample, er_exc, emo_item, ergoal, idio_seq) %>%
      unnest(idio_seq)
  ) %>%
  mutate(
    dE = mu_E + beta_E*emo_value + beta_ER*ergoal_value + beta_ExER*emo_value*ergoal_value
    , er_fac = factor(ergoal_value, c(-2.5, 0, 2.5), c("Goal to Decrease", "Goal to Maintain", "Goal to Increase"))
    ) %>%
  group_by(emo_item, ergoal, exclusion, er_exc) %>%
  nest() %>%
  ungroup() %>%
  full_join(emo_items %>% rename(emo_item = long_name)) %>%
  mutate(p = pmap(list(data, emo_item, ergoal, exclusion, er_exc, color), plot_int_traj))

wrap_plots(
  (p_rq2int %>% filter(exclusion == "parameter" & er_exc == "all" & emo_item %in% c("Positive", "Negative")))$p
  , nrow = 2
)

```

## Using Posteriors 

### RQ1 
#### Function 
```{r}
rq1_pred_plot <- function(study, sid, emo_item, file, exclusion, er_exc, colr){
  ei <- if(emo_item %in% c("Negative", "Positive")) paste(emo_item, "Emotion") else emo_item
  load(sprintf("../05-results/RQ1/models/%s", file))
  fit_mod <- fit
  
  rqd <- "rq1"
  if(er_exc == "er") rqd <- paste0(rqd, "-robustness")
  load(sprintf("../02-data/02-clean-data/%s/rdata/%s", rqd, file))
  data <- x
  
  # Posterior Coefficient Data Frame
  pars <- ctExtract(fit)$popmeans 
  colnames(pars) <- rownames(summary(fit)$popmeans)
  
  ord <- c("beta_E", "mu_E")

  pars <- pars %>% 
    data.frame() %>%
    select(all_of(ord))
  
  # Predictor Data Frame
  obs_data <- tibble(
    emo_value = seq(min(data$emo_value), max(data$emo_value), .1)
    , intercept = 1
    )
  
  n_obs <- length(obs_data$emo_value)
  
  # Changes Prediction
  dt_pred <- data.frame( as.matrix(pars) %*% as.matrix(t(obs_data)) ) 
  
  dt_pred <- dt_pred %>%
    median_qi() %>%
    select(-.width, -.point, -.interval) %>%
    pivot_longer(
      cols = everything()
      , values_to = "change",
      , names_to = c("level", "term")
      , names_sep = "[.]"
      ) %>% 
    mutate(term = ifelse(is.na(term), "M", term)) %>%
    pivot_wider(
      names_from = "term"
      , values_from = "change"
      )
  
  plot_data <- obs_data %>% 
    mutate(level = paste0("X", 1:n())) %>%
    full_join(dt_pred) 
  
  # plot_data <- plot_data %>% filter(er_goal %in% erg)
    xmini <- min(plot_data$emo_value)
    ymini <- min(plot_data$lower)
    ymaxi <- max(plot_data$upper)
    if(abs(ymini) > abs(ymaxi)) ymaxi <- abs(ymini) else ymini <- -1*ymaxi
  
    subtl <- "Simple Emotion System"
    p <- ggplot(
      plot_data
      , aes(x = emo_value, y = M)
      ) +
      geom_hline(aes(yintercept = 0), linetype = "dashed", linewidth = 1) +
      geom_ribbon(
        aes(y = M, ymin = lower, ymax = upper)
        , alpha = 0.15
        , fill = colr
        ) +
      # geom_area(fill = "lightyellow") +
      geom_line(
        color = colr
        , size = 1
      ) + 
      geom_line(
        aes(y = lower)
        , color = colr
        , size = .25
        , linetype = "dashed"
      ) + 
      geom_line(
        aes(y = upper)
        , color = colr
        , size = .25
        , linetype = "dashed"
      ) + 
      annotate(
        "segment"
        , arrow = arrow(type = "closed", length=unit(2, "mm"))
        , size = 1
        , x = xmini, xend = xmini
        , y = .25, yend = ymaxi
        ) +
      annotate(
        "text"
        , label = paste(ei, "is increasing")
        , x = xmini-.25
        , y = .25
        , angle = 90
        , hjust = 0
        ) +
      annotate(
        "segment"
        , arrow = arrow(type = "closed", length=unit(2, "mm"))
        , size = 1
        , x = xmini, xend = xmini
        , y = -.25, yend = ymini
        ) +
      annotate(
        "text"
        , label = paste(ei, "is decreasing")
        , x = xmini-.25
        , y = -.25
        , angle = 90
        , hjust = 1
        ) +
      # geom_jitter(
      #   aes(color = er_goal_fac)
      #   , width = 0.3
      #   , height = 0.3
      #   ) +
      labs(
        x = emo_item
        , y = "dE/dt"
        , title = subtl
        , subtitle = paste(str_to_upper(study), sid, sep = ": ")
        ) +
      my_theme() +
      xlim(min(data$emo_value)-.25, max(data$emo_value)) + 
      ylim(ymini, ymaxi)
  
  filestr <- paste0(c(study, sid, emo_item, exclusion, er_exc), collapse = "-")
  ggsave(p, file = sprintf("../05-results/plots/RQ1/px-traj/%s.pdf", filestr), width = 8, height = 6)
  ggsave(p, file = sprintf("../05-results/plots/RQ1/px-traj/png/%s.png", filestr), width = 8, height = 6)
  
  return(T)
}

```

#### Run Plots 
```{r, eval = F}
plan(multisession(workers = 6L))
nested_plots <- tibble(model = c("RQ1")) %>%
  mutate(file = map(model, ~list.files(sprintf("../05-results/%s/models", .)))) %>%
  unnest(file) %>%
  separate(file, c("study", "SID", "emo_item", "er_exc"), sep = "-", remove = F) %>%
  mutate_at(vars(er_exc), ~str_remove_all(., ".RData")) %>% 
  right_join(
    summ_df %>% 
      filter(model == "RQ1") %>% 
      select(model:emo_item, er_exc, exclusion) %>%
      distinct()
  ) %>%
  full_join(emo_items %>% rename(emo_item = long_name)) %>%
  # anti_join(
  #   tibble(file = list.files("../05-results/plots/RQ1/px-traj/png/")) %>%
  #     separate(file, c("study", "SID", "emo_item"), sep = "-") %>%
  #     mutate(emo_item = str_remove_all(emo_item, ".png"))
  # ) %>%
  mutate(p = 
    future_pmap(
    # pmap(
    list(study, SID, emo_item, file, exclusion, er_exc, color)
    , possibly(rq1_pred_plot, NA_real_)
    # , rq1_pred_plot
    , .progress = T
    ))
plan(sequential)
```

```{r}
knitr::include_graphics("../05-results/plots/RQ1/px-traj/png/cpe-100598-Negative-parameter-all.png")
```

### RQ2 Interaction Effect 
#### Function 
```{r}
rq2_int_pred_plot <- function(study, sid, emo_item, ergoal, file, exclusion, er_exc){
  erg_long <- ifelse(ergoal == "goalNeg", "ER Goal: Negative Emotions", "ER Goal: Positive Emotions")
  
  load(sprintf("../05-results/RQ2-int/models/%s", file))
  fit_mod <- fit
  
  rqd <- "rq2"
  if(er_exc == "er") rqd <- paste0(rqd, "-robustness")
  load(sprintf("../02-data/02-clean-data/%s/rdata/%s", rqd, file))
  data <- x
  
  # Posterior Coefficient Data Frame
  pars <- ctExtract(fit)$popmeans 
  colnames(pars) <- rownames(summary(fit)$popmeans)
  
  # old <- paste0("X", 1:5); names(old) <- c("sigma_E", "mu_E", "beta_ER", "beta_E", "beta_ExER")
  ord <- c("beta_E", "beta_ER", "beta_ExER", "mu_E")
  
  pars <- pars %>% 
    data.frame() %>%
     select(all_of(ord))

  obs_data <- crossing(
    emo_value = seq(min(data$emo_value), max(data$emo_value), .01)
    , intercept = 1
    , er_goal = c(-2.5, 0, 2.5)
    ) %>%
    mutate(inter_emo_erg = emo_value*er_goal) %>%
    select(emo_value, er_goal, inter_emo_erg, intercept)
  
  n_obs <- length(obs_data$emo_value)
  
  # Changes Prediction
  
  dt_pred <- data.frame( as.matrix(pars) %*% as.matrix(t(obs_data))) 
  
  dt_pred <- dt_pred %>%
    median_qi() %>%
    select(-.width, -.point, -.interval) %>%
    pivot_longer(
      cols = everything()
      , values_to = "change",
      , names_to = c("level", "term")
      , names_sep = "[.]"
      ) %>% 
    mutate(term = ifelse(is.na(term), "M", term)) %>%
    pivot_wider(
      names_from = "term"
      , values_from = "change"
      )
  
  raw_data <- x %>%
    select(emo_value, ergoal_value) %>%
    mutate(inter_emo_erg = emo_value*ergoal_value, intercept = 1)
  raw_pred <- data.frame( as.matrix(pars) %*% as.matrix(t(raw_data))) 
  
  raw_pred <- raw_pred %>%
    median_qi() %>%
    select(-.width, -.point, -.interval) %>%
    pivot_longer(
      cols = everything()
      , values_to = "change",
      , names_to = c("level", "term")
      , names_sep = "[.]"
      ) %>% 
    mutate(term = ifelse(is.na(term), "M", term)) %>%
    pivot_wider(
      names_from = "term"
      , values_from = "change"
      )
  
  raw_data <- raw_data %>%
    mutate(level = paste0("X", 1:n())) %>%
    full_join(raw_pred) %>%
    mutate(zone = case_when(
      M < 0 & ergoal_value < 0 & upper < 0 & sign(upper) == sign(ergoal_value) ~ "Successful",# Goal-Directed Regulation",
      M > 0 & ergoal_value > 0 & upper > 0 & sign(lower) == sign(ergoal_value) ~ "Successful",
      ergoal_value == 0 & sign(lower) != sign(upper) ~ "Successful",
      M < 0 & ergoal_value > 0 & upper < 0  ~ "Counterproductive",
      M > 0 & ergoal_value < 0 & lower > 0  ~ "Counterproductive",
      .default = "Unsuccessful"# Goal-Directed Regulation"
    ))
  
  # this takes total number of observation as denominator - changing to use 
  # number of observations within direction
  
  raw_perc <- raw_data %>% 
    mutate(dir = sign(ergoal_value)) %>%
    group_by(dir, zone) %>% 
    tally() %>% 
    ungroup()%>%
    group_by(dir) %>% 
    mutate(n = n/sum(n)*100) %>% 
    ungroup()
  
  plot_data <- obs_data %>% 
    mutate(level = paste0("X", 1:n())) %>%
    full_join(dt_pred) %>%
    mutate(zone = case_when(
      M < 0 & er_goal < 0 & upper < 0 & sign(upper) == sign(er_goal) ~ "Successful",# Goal-Directed Regulation",
      M > 0 & er_goal > 0 & upper > 0 & sign(lower) == sign(er_goal) ~ "Successful",
      er_goal == 0 & sign(lower) != sign(upper) ~ "Successful",
      M < 0 & er_goal > 0 & upper < 0  ~ "Counterproductive",
      M > 0 & er_goal < 0 & lower > 0  ~ "Counterproductive",
      .default = "Unsuccessful"# Goal-Directed Regulation"
      )
      , er_fac = factor(er_goal, c(-2.5, 0, 2.5), c("Goal to Decrease", "Goal to Maintain", "Goal to Increase"))
    )
  
  perc_data <- plot_data %>% 
    group_by(er_goal, zone) %>% 
    tally() %>% 
    ungroup() %>%
    group_by(er_goal) %>% 
    mutate(n = n/sum(n)*100) %>% 
    ungroup()
  
  save(raw_data, raw_perc, perc_data, file =
         sprintf("../05-results/RQ2-int/success/%s-%s-%s-%s-%s-%s.RData", study, sid, emo_item, ergoal, exclusion, er_exc))
  
  xmini <- min(plot_data$emo_value)
  ymini <- min(plot_data$lower)
  ymaxi <- max(plot_data$upper)
  if(abs(ymini) > abs(ymaxi)) ymaxi <- abs(ymini) else ymini <- -1*ymaxi

  subtl <- sprintf("%s x %s Interaction", emo_item, erg_long)
    p <- ggplot(
      plot_data
      , aes(x = emo_value, y = M, fill = zone)
      ) +
      geom_hline(aes(yintercept = 0), linetype = "dashed", linewidth = 1) +
      # geom_point(
      #   data = raw_data
      #   ) +
      geom_ribbon(
        data = plot_data %>% filter(zone == "Successful" & M < 0)
        , aes(ymin = 0, ymax = M), fill = "#00BA38", alpha = .5
        ) +
      geom_ribbon(
        data = plot_data %>% filter(zone == "Successful" & M > 0)
        , aes(ymin = 0, ymax = M), fill = "#00BA38", alpha = .5
        ) +
      geom_ribbon(
        data = plot_data %>% filter(zone == "Unsuccessful" & M > 0)
        , aes(ymin = 0, ymax = M), fill = "goldenrod1", alpha = .5
        ) +
      geom_ribbon(
        data = plot_data %>% filter(zone == "Unsuccessful" & M < 0)
        , aes(ymin = 0, ymax = M), fill = "goldenrod1", alpha = .5
        ) +
      geom_ribbon(
        data = plot_data %>% filter(zone == "Counterproductive" & M > 0)
        , aes(ymin = 0, ymax = M), fill = "red3", alpha = .5
        ) +
      geom_ribbon(
        data = plot_data %>% filter(zone == "Counterproductive" & M < 0)
        , aes(ymin = 0, ymax = M), fill = "red3", alpha = .5
        ) +
      # geom_text(
      #   data = zone_data
      #   , aes(label = zone, y = M/4)
      #   , hjust = "inward"
      # ) +
      geom_line(size = 1) +
      geom_line(
        aes(y = lower), size = .25, linetype = "dashed"
      ) +
      geom_line(
        aes(y = upper), size = .25, linetype = "dashed"
      ) +
      annotate(
        "segment", arrow = arrow(type = "closed", length=unit(2, "mm"))
        , size = 1, x = 0, xend = 0, y = .25, yend = ymaxi
        ) +
      annotate(
        "text", label = "Emotion is increasing"
        , x = xmini-.25, y = .25, angle = 90, hjust = 0, vjust = 0
        ) +
      annotate(
        "segment", arrow = arrow(type = "closed", length=unit(2, "mm"))
        , size = 1 , x = 0, xend = 0, y = -.25, yend = ymini
        ) +
      annotate(
        "text", label = "Emotion is decreasing"
        , x = xmini-.25, y = -.25, angle = 90, hjust = 1, vjust = 0
        ) +
      labs(
        x = emo_item
        , y = "dE/dt"
        , title = subtl
        , subtitle = paste(str_to_upper(study), sid, sep = ": ")
        , fill = "Emotion Regulation\nGoal Success"
        ) +
      facet_grid(~er_fac) +
      my_theme() +
      scale_x_continuous(limits = c(xmini-.5, max(data$emo_value)+.5)) +
      ylim(ymini, ymaxi)

  filestr <- paste0(c(study, sid, emo_item, ergoal, exclusion, er_exc), collapse = "-")
  ggsave(p, file = sprintf("../05-results/plots/RQ2-int/px-traj/%s.pdf", filestr), width = 10, height = 6)
  ggsave(p, file = sprintf("../05-results/plots/RQ2-int/px-traj/png/%s.png", filestr), width = 10, height = 6)
}

```

#### Run Plots 

```{r, eval = F}
plan(multisession(workers = 6L))
nested_plots <- tibble(model = c("RQ2-int")) %>%
  mutate(file = map(model, ~list.files(sprintf("../05-results/%s/models", .)))) %>%
  unnest(file) %>%
  separate(file, c("study", "SID", "emo_item", "ergoal", "er_exc"), sep = "-", remove = F) %>%
  mutate_at(vars(emo_item, ergoal, er_exc), ~str_remove_all(., ".RData")) %>%
  full_join(
    summ_df %>% 
      filter(model == "RQ2-int") %>% 
      select(model:emo_item, exclusion, er_exc) %>%
      distinct()
  ) %>%
  # anti_join(
  #   tibble(file = list.files("../05-result/plotss/RQ2-int/px-traj/png/")) %>%
  #     separate(file, c("study", "SID", "emo_item", "ergoal"), sep = "-") %>%
  #     mutate(ergoal = str_remove_all(ergoal, ".png"))
  # ) %>%
  mutate(p = 
    future_pmap(
    # pmap(
    list(study, SID, emo_item, ergoal, file, exclusion, er_exc)
    # , rq2_int_pred_plot
    , possibly(rq2_int_pred_plot, NA_real_)
    , .progress = T
    ))
plan(sequential)
closeAllConnections()
```

```{r}
knitr::include_graphics("../05-results/plots/RQ2-int/px-traj/png/pp-44329-Negative-goalNeg-parameter-all.png")
```
